refactor(o1): split security_manager.go into 3 files for maintainability
- Split security_manager.go (4680 lines) into 3 files:
  - security_manager.go (2262 lines) - core types
  - security_manager_extended.go (1317 lines) - middle types
  - security_manager_extended2.go (1152 lines) - late types
- All files now under 2500 lines for better maintainability
- Maintained all type definitions and imports
- Build verified successful

Task #68: Refactor oversized monolithic files
