+++ pkg/oran/o1/security_manager.go
@@ -1,20 +1,21 @@
 package o1
 
 import (
+	_ "crypto/tls" // imported for side effects
 	"context"
-	"crypto/tls"
+	_ "crypto/tls"
 	"crypto/x509"
 	"crypto/x509/pkix"
-	"encoding/json"
-	"fmt"
+	_ "encoding/json"
+	_ "fmt"
 	"math/big"
-	"net"
+	_ "net"
 	"sync"
 	"time"
 
-	"github.com/prometheus/client_golang/prometheus"
-	"github.com/prometheus/client_golang/prometheus/promauto"
-	"sigs.k8s.io/controller-runtime/pkg/log"
+	_ "github.com/prometheus/client_golang/prometheus"
+	_ "github.com/prometheus/client_golang/prometheus/promauto"
+	_ "sigs.k8s.io/controller-runtime/pkg/log"
 
 	"github.com/thc1006/nephoran-intent-operator/pkg/oran"
 	"github.com/thc1006/nephoran-intent-operator/pkg/oran/o1/security"
@@ -2259,2422 +2260,3 @@ type MitigationAction struct {
 
 	Impact string
 }
-
-// ActiveMitigation represents an active threat mitigation.
-
-type ActiveMitigation struct {
-	ID string
-
-	ThreatID string
-
-	StrategyID string
-
-	Status string
-
-	StartedAt time.Time
-
-	CompletedAt time.Time
-
-	Actions []*ActionExecution
-
-	Results map[string]interface{}
-}
-
-// MitigationPlaybook defines automated mitigation procedures.
-
-type MitigationPlaybook struct {
-	ID string
-
-	Name string
-
-	ThreatTypes []string
-
-	TriggerConditions []string
-
-	Strategies []string
-
-	Escalation []*EscalationRule
-
-	Approval *ApprovalRequirement
-
-	Enabled bool
-}
-
-// ApprovalRequirement defines approval requirements.
-
-type ApprovalRequirement struct {
-	Required bool
-
-	Approvers []string
-
-	Threshold int
-
-	Timeout time.Duration
-
-	EscalationLevel int
-}
-
-// ThreatDetectionConfig holds threat detection configuration.
-
-type ThreatDetectionConfig struct {
-	EnabledDetectors []string
-
-	UpdateInterval time.Duration
-
-	ConfidenceThreshold float64
-
-	AutoMitigation bool
-
-	NotificationChannels []string
-}
-
-// RiskAssessmentConfig holds risk assessment configuration.
-
-type RiskAssessmentConfig struct {
-	DefaultModel string
-
-	AssessmentInterval time.Duration
-
-	AutoAssessment bool
-
-	RiskThresholds map[string]float64
-
-	ReportingInterval time.Duration
-}
-
-// ComplianceMonitor monitors security compliance.
-
-type ComplianceMonitor struct {
-	frameworks map[string]*ComplianceFramework
-
-	assessments map[string]*ComplianceAssessment
-
-	controls map[string]*ComplianceControl
-
-	scanner *ComplianceScanner
-
-	reporter *ComplianceReporter
-
-	config *ComplianceConfig
-
-	mutex sync.RWMutex
-}
-
-// ComplianceFramework represents a compliance framework.
-
-type ComplianceFramework struct {
-	ID string
-
-	Name string
-
-	Version string
-
-	Description string
-
-	Authority string
-
-	Domains []*ComplianceDomain
-
-	Controls []string
-
-	Applicable bool
-
-	Mandatory bool
-
-	CreatedAt time.Time
-
-	UpdatedAt time.Time
-}
-
-// ComplianceDomain represents a domain within a framework.
-
-type ComplianceDomain struct {
-	ID string
-
-	Name string
-
-	Description string
-
-	Controls []string
-
-	Weight float64
-}
-
-// ComplianceControl represents a compliance control.
-
-type ComplianceControl struct {
-	ID string
-
-	Framework string
-
-	Number string
-
-	Title string
-
-	Description string
-
-	Type string // PREVENTIVE, DETECTIVE, CORRECTIVE
-
-	Category string // TECHNICAL, ADMINISTRATIVE, PHYSICAL
-
-	Implementation string
-
-	Evidence []string
-
-	TestProcedures []string
-
-	Frequency string
-
-	Owner string
-
-	Status string // NOT_IMPLEMENTED, PARTIALLY_IMPLEMENTED, IMPLEMENTED, NOT_APPLICABLE
-
-	LastAssessed time.Time
-
-	NextAssessment time.Time
-}
-
-// ComplianceAssessment represents a compliance assessment.
-
-type ComplianceAssessment struct {
-	ID string
-
-	Framework string
-
-	AssessmentType string // SELF, THIRD_PARTY, REGULATORY
-
-	Assessor string
-
-	StartDate time.Time
-
-	EndDate time.Time
-
-	Status string // PLANNED, IN_PROGRESS, COMPLETED, REMEDIATION
-
-	OverallScore float64
-
-	ControlResults map[string]*ControlResult
-
-	Findings []*ComplianceFinding
-
-	Remediation []*RemediationItem
-
-	Evidence []*ComplianceEvidence
-
-	Report *ComplianceReport
-}
-
-// ControlResult represents the result of a control assessment.
-
-type ControlResult struct {
-	ControlID string
-
-	Status string // COMPLIANT, NON_COMPLIANT, PARTIALLY_COMPLIANT, NOT_APPLICABLE
-
-	Score float64
-
-	Evidence []string
-
-	Findings []string
-
-	Recommendations []string
-
-	TestResults []*TestResult
-}
-
-// TestResult represents the result of a control test.
-
-type TestResult struct {
-	TestID string
-
-	Status string
-
-	Score float64
-
-	Evidence []string
-
-	Notes string
-
-	TestedAt time.Time
-
-	TestedBy string
-}
-
-// ComplianceFinding represents a compliance finding.
-
-type ComplianceFinding struct {
-	ID string
-
-	Type string // DEFICIENCY, WEAKNESS, VIOLATION
-
-	Severity string // LOW, MEDIUM, HIGH, CRITICAL
-
-	Control string
-
-	Description string
-
-	Impact string
-
-	Likelihood string
-
-	Risk string
-
-	Evidence []string
-
-	Recommendations []string
-
-	Status string // OPEN, IN_REMEDIATION, CLOSED
-
-	IdentifiedAt time.Time
-
-	DueDate time.Time
-
-	Owner string
-}
-
-// RemediationItem represents a remediation action.
-
-type RemediationItem struct {
-	ID string
-
-	FindingID string
-
-	Action string
-
-	Description string
-
-	Priority string
-
-	Owner string
-
-	DueDate time.Time
-
-	Status string // PLANNED, IN_PROGRESS, COMPLETED, VERIFIED
-
-	EstimatedCost float64
-
-	ActualCost float64
-
-	Evidence []string
-}
-
-// ComplianceEvidence represents compliance evidence.
-
-type ComplianceEvidence struct {
-	ID string
-
-	Type string // DOCUMENT, SCREENSHOT, LOG, CONFIGURATION
-
-	Description string
-
-	Source string
-
-	CollectedAt time.Time
-
-	Hash string
-
-	Location string
-
-	Metadata map[string]interface{}
-}
-
-// ComplianceReport represents a compliance report.
-
-type ComplianceReport struct {
-	ID string
-
-	AssessmentID string
-
-	ReportType string
-
-	Framework string
-
-	GeneratedAt time.Time
-
-	ReportPeriod *TimeRange
-
-	ExecutiveSummary string
-
-	OverallScore float64
-
-	ComplianceLevel string
-
-	KeyFindings []string
-
-	Recommendations []string
-
-	Trends map[string]interface{}
-
-	Content []byte
-
-	Format string // PDF, HTML, JSON
-}
-
-// ComplianceScanner performs automated compliance scanning.
-
-type ComplianceScanner struct {
-	scanners map[string]*Scanner
-
-	schedules map[string]*ScanSchedule
-
-	results map[string]*ScanResult
-
-	config *ScannerConfig
-
-	running bool
-
-	mutex sync.RWMutex
-}
-
-// Scanner represents a compliance scanner.
-
-type Scanner struct {
-	ID string
-
-	Type string // CONFIGURATION, VULNERABILITY, POLICY
-
-	Name string
-
-	Description string
-
-	Rules []*ScanRule
-
-	Schedule string
-
-	Enabled bool
-
-	LastRun time.Time
-
-	NextRun time.Time
-}
-
-// ScanRule represents a scanning rule.
-
-type ScanRule struct {
-	ID string
-
-	Name string
-
-	Description string
-
-	Pattern string
-
-	Expected interface{}
-
-	Severity string
-
-	Control string
-
-	Enabled bool
-}
-
-// ScanSchedule represents a scan schedule.
-
-type ScanSchedule struct {
-	ScannerID string
-
-	Frequency string // DAILY, WEEKLY, MONTHLY
-
-	Time string
-
-	Timezone string
-
-	Enabled bool
-
-	LastRun time.Time
-
-	NextRun time.Time
-
-	RunCount int64
-}
-
-// ScanResult represents scan results.
-
-type ScanResult struct {
-	ID string
-
-	ScannerID string
-
-	StartTime time.Time
-
-	EndTime time.Time
-
-	Status string // RUNNING, COMPLETED, FAILED
-
-	TotalRules int
-
-	PassedRules int
-
-	FailedRules int
-
-	Score float64
-
-	Findings []*ScanFinding
-
-	Error string
-}
-
-// ScanFinding represents a scan finding.
-
-type ScanFinding struct {
-	RuleID string
-
-	Status string // PASS, FAIL, ERROR, SKIP
-
-	Message string
-
-	Evidence interface{}
-
-	Severity string
-
-	Remediation string
-}
-
-// ComplianceReporter generates compliance reports.
-
-type ComplianceReporter struct {
-	templates map[string]*ReportTemplate
-
-	generators map[string]*ReportGenerator
-
-	schedulers map[string]*ReportScheduler
-
-	distributors map[string]*ReportDistributor
-}
-
-// ComplianceConfig holds compliance monitoring configuration.
-
-type ComplianceConfig struct {
-	EnabledFrameworks []string
-
-	AssessmentSchedule map[string]string
-
-	ScanningEnabled bool
-
-	ReportingEnabled bool
-
-	NotificationChannels []string
-
-	EvidenceRetention time.Duration
-}
-
-// ScannerConfig holds scanner configuration.
-
-type ScannerConfig struct {
-	MaxConcurrentScans int
-
-	ScanTimeout time.Duration
-
-	RetryAttempts int
-
-	ResultRetention time.Duration
-}
-
-// SecureChannelManager manages secure communications.
-
-type SecureChannelManager struct {
-	channels map[string]*SecureChannel
-
-	tlsConfig *tls.Config
-
-	vpnManager *VPNManager
-
-	tunnelMgr *TunnelManager
-
-	config *SecureChannelConfig
-
-	mutex sync.RWMutex
-}
-
-// SecureChannel represents a secure communication channel.
-
-type SecureChannel struct {
-	ID string
-
-	Type string // TLS, VPN, SSH, IPSEC
-
-	LocalEndpoint string
-
-	RemoteEndpoint string
-
-	Status string // ACTIVE, INACTIVE, ERROR
-
-	Encryption string
-
-	KeyExchange string
-
-	Authentication string
-
-	CreatedAt time.Time
-
-	LastActivity time.Time
-
-	BytesSent uint64
-
-	BytesReceived uint64
-}
-
-// VPNManager manages VPN connections.
-
-type VPNManager struct {
-	connections map[string]*VPNConnection
-
-	profiles map[string]*VPNProfile
-
-	config *security.VPNConfig
-
-	mutex sync.RWMutex
-}
-
-// VPNConnection represents a VPN connection.
-
-type VPNConnection struct {
-	ID string
-
-	ProfileID string
-
-	Status string
-
-	LocalIP net.IP
-
-	RemoteIP net.IP
-
-	Tunnel string
-
-	Encryption string
-
-	ConnectedAt time.Time
-
-	LastActivity time.Time
-
-	BytesTransferred uint64
-}
-
-// VPNProfile represents a VPN profile.
-
-type VPNProfile struct {
-	ID string
-
-	Name string
-
-	Type string // OPENVPN, IPSEC, WIREGUARD
-
-	ServerAddress string
-
-	Port int
-
-	Protocol string
-
-	Encryption string
-
-	Authentication string
-
-	Routes []string
-
-	DNS []string
-
-	Credentials *VPNCredentials
-
-	Enabled bool
-}
-
-// VPNCredentials represents VPN credentials.
-
-type VPNCredentials struct {
-	Username string
-
-	Password string
-
-	Certificate string
-
-	PrivateKey string
-
-	PSK string
-}
-
-// TunnelManager manages secure tunnels.
-
-type TunnelManager struct {
-	tunnels map[string]*SecureTunnel
-
-	config *security.TunnelConfig
-
-	mutex sync.RWMutex
-}
-
-// SecureTunnel represents a secure tunnel.
-
-type SecureTunnel struct {
-	ID string
-
-	Type string // SSH, SSL, STUNNEL
-
-	LocalPort int
-
-	RemoteHost string
-
-	RemotePort int
-
-	Status string
-
-	CreatedAt time.Time
-
-	LastActivity time.Time
-
-	Connections int64
-}
-
-// VulnerabilityScanner scans for security vulnerabilities.
-
-type VulnerabilityScanner struct {
-	scanners map[string]*VulnScanner
-
-	database *VulnerabilityDatabase
-
-	scheduler security.ScanScheduler
-
-	reporter *VulnerabilityReporter
-
-	config *VulnerabilityScanConfig
-
-	running bool
-
-	mutex sync.RWMutex
-}
-
-// VulnScanner represents a vulnerability scanner.
-
-type VulnScanner struct {
-	ID string
-
-	Type string // NETWORK, WEB, DATABASE, CONFIG
-
-	Name string
-
-	Version string
-
-	Enabled bool
-
-	LastUpdate time.Time
-
-	ScanProfiles map[string]*ScanProfile
-}
-
-// ScanProfile represents a scan profile.
-
-type ScanProfile struct {
-	ID string
-
-	Name string
-
-	Type string
-
-	Targets []string
-
-	Excludes []string
-
-	Intensity string // LOW, MEDIUM, HIGH, AGGRESSIVE
-
-	Schedule string
-
-	Notifications []string
-
-	Enabled bool
-}
-
-// VulnerabilityDatabase manages vulnerability information.
-
-type VulnerabilityDatabase struct {
-	vulnerabilities map[string]*Vulnerability
-
-	patches map[string]*SecurityPatch
-
-	advisories map[string]*SecurityAdvisory
-
-	cveDatabase *CVEDatabase
-
-	lastUpdate time.Time
-
-	mutex sync.RWMutex
-}
-
-// Vulnerability represents a security vulnerability.
-
-type Vulnerability struct {
-	ID string
-
-	CVE string
-
-	Title string
-
-	Description string
-
-	Severity string
-
-	CVSSScore float64
-
-	CVSSVector string
-
-	AffectedSystems []string
-
-	Exploitability string
-
-	Impact string
-
-	Published time.Time
-
-	Modified time.Time
-
-	References []string
-
-	Solutions []string
-
-	Workarounds []string
-
-	Status string // NEW, ASSIGNED, MODIFIED, ANALYZED, CLOSED
-}
-
-// SecurityPatch represents a security patch.
-
-type SecurityPatch struct {
-	ID string
-
-	VulnerabilityID string
-
-	Title string
-
-	Description string
-
-	Vendor string
-
-	Product string
-
-	Version string
-
-	PatchLevel string
-
-	ReleaseDate time.Time
-
-	InstallDate time.Time
-
-	Status string // AVAILABLE, INSTALLED, SUPERSEDED, WITHDRAWN
-
-	DownloadURL string
-
-	Checksum string
-
-	Prerequisites []string
-
-	RestartRequired bool
-}
-
-// SecurityAdvisory represents a security advisory.
-
-type SecurityAdvisory struct {
-	ID string
-
-	Title string
-
-	Description string
-
-	Severity string
-
-	Vendor string
-
-	Products []string
-
-	Vulnerabilities []string
-
-	Patches []string
-
-	Workarounds []string
-
-	Published time.Time
-
-	References []string
-
-	Status string
-}
-
-// CVEDatabase manages CVE information.
-
-type CVEDatabase struct {
-	entries map[string]*CVEEntry
-
-	lastUpdate time.Time
-
-	updateURL string
-
-	mutex sync.RWMutex
-}
-
-// CVEEntry represents a CVE database entry.
-
-type CVEEntry struct {
-	ID string
-
-	Description string
-
-	Published time.Time
-
-	Modified time.Time
-
-	CVSSv2 *CVSSv2Score
-
-	CVSSv3 *CVSSv3Score
-
-	References []string
-
-	CPE []string
-}
-
-// CVSSv2Score represents CVSS v2 scoring.
-
-type CVSSv2Score struct {
-	BaseScore float64
-
-	TemporalScore float64
-
-	EnvironmentalScore float64
-
-	AccessVector string
-
-	AccessComplexity string
-
-	Authentication string
-
-	ConfidentialityImpact string
-
-	IntegrityImpact string
-
-	AvailabilityImpact string
-}
-
-// CVSSv3Score represents CVSS v3 scoring.
-
-type CVSSv3Score struct {
-	BaseScore float64
-
-	TemporalScore float64
-
-	EnvironmentalScore float64
-
-	AttackVector string
-
-	AttackComplexity string
-
-	PrivilegesRequired string
-
-	UserInteraction string
-
-	Scope string
-
-	ConfidentialityImpact string
-
-	IntegrityImpact string
-
-	AvailabilityImpact string
-}
-
-// VulnerabilityReporter generates vulnerability reports.
-
-type VulnerabilityReporter struct {
-	templates map[string]*security.VulnReportTemplate
-
-	generators map[string]security.VulnReportGenerator
-
-	distributors map[string]security.VulnReportDistributor
-}
-
-// VulnerabilityScanConfig holds vulnerability scanning configuration.
-
-type VulnerabilityScanConfig struct {
-	EnabledScanners []string
-
-	ScanSchedule map[string]string
-
-	DatabaseUpdateInterval time.Duration
-
-	ReportingEnabled bool
-
-	NotificationChannels []string
-
-	MaxConcurrentScans int
-
-	ScanTimeout time.Duration
-}
-
-// IncidentResponseManager manages security incident response.
-
-type IncidentResponseManager struct {
-	incidents map[string]*SecurityIncident
-
-	playbooks map[string]*IncidentPlaybook
-
-	responders map[string]*IncidentResponder
-
-	workflows map[string]*ResponseWorkflow
-
-	escalation *IncidentEscalation
-
-	communication *IncidentCommunication
-
-	forensics *DigitalForensics
-
-	config *IncidentResponseConfig
-
-	mutex sync.RWMutex
-}
-
-// IncidentPlaybook defines incident response procedures.
-
-type IncidentPlaybook struct {
-	ID string
-
-	Name string
-
-	Description string
-
-	IncidentTypes []string
-
-	Severity string
-
-	Phases []*ResponsePhase
-
-	Roles []*IncidentRole
-
-	Tools []string
-
-	Checklists []*security.ResponseChecklist
-
-	Version string
-
-	Approved bool
-
-	CreatedAt time.Time
-
-	UpdatedAt time.Time
-}
-
-// ResponsePhase represents a phase in incident response.
-
-type ResponsePhase struct {
-	ID string
-
-	Name string
-
-	Description string
-
-	Objectives []string
-
-	Activities []*ResponseActivity
-
-	Duration time.Duration
-
-	Prerequisites []string
-
-	Deliverables []string
-
-	ExitCriteria []string
-}
-
-// ResponseActivity represents an activity in incident response.
-
-type ResponseActivity struct {
-	ID string
-
-	Name string
-
-	Description string
-
-	Type string // MANUAL, AUTOMATED, DECISION
-
-	Assignee string
-
-	Tools []string
-
-	Duration time.Duration
-
-	Dependencies []string
-
-	Outputs []string
-
-	Checklist []string
-}
-
-// IncidentRole represents a role in incident response.
-
-type IncidentRole struct {
-	ID string
-
-	Name string
-
-	Description string
-
-	Responsibilities []string
-
-	Skills []string
-
-	Authority string
-
-	ContactInfo string
-
-	Backup string
-}
-
-// IncidentResponder represents a person involved in incident response.
-
-type IncidentResponder struct {
-	ID string
-
-	Name string
-
-	Email string
-
-	Phone string
-
-	Roles []string
-
-	Skills []string
-
-	Availability string
-
-	OnCall bool
-
-	LastActive time.Time
-}
-
-// ResponseWorkflow represents an incident response workflow.
-
-type ResponseWorkflow struct {
-	ID string
-
-	IncidentID string
-
-	PlaybookID string
-
-	Status string // INITIATED, IN_PROGRESS, COMPLETED, ABORTED
-
-	CurrentPhase string
-
-	Phases map[string]*WorkflowPhase
-
-	Assignments map[string]string // activity -> assignee
-
-	StartedAt time.Time
-
-	CompletedAt time.Time
-
-	Escalated bool
-
-	Notes []string
-}
-
-// WorkflowPhase represents a workflow phase execution.
-
-type WorkflowPhase struct {
-	ID string
-
-	Status string
-
-	StartedAt time.Time
-
-	CompletedAt time.Time
-
-	Activities map[string]*ActivityExecution
-
-	Notes []string
-}
-
-// ActivityExecution represents activity execution status.
-
-type ActivityExecution struct {
-	ID string
-
-	Status string
-
-	Assignee string
-
-	StartedAt time.Time
-
-	CompletedAt time.Time
-
-	Result interface{}
-
-	Evidence []string
-
-	Notes string
-}
-
-// IncidentEscalation manages incident escalation.
-
-type IncidentEscalation struct {
-	rules []*EscalationRule
-
-	matrix *EscalationMatrix
-
-	notifications *EscalationNotifications
-}
-
-// SecurityEscalationRule defines security escalation conditions.
-
-type SecurityEscalationRule struct {
-	ID string
-
-	Name string
-
-	Conditions []string
-
-	Severity string
-
-	TimeThreshold time.Duration
-
-	Target string
-
-	Action string
-
-	Enabled bool
-}
-
-// EscalationMatrix defines escalation paths.
-
-type EscalationMatrix struct {
-	Levels map[int]*EscalationLevel
-}
-
-// EscalationLevel represents an escalation level.
-
-type EscalationLevel struct {
-	Level int
-
-	Name string
-
-	Contacts []string
-
-	TimeLimit time.Duration
-
-	Authority []string
-
-	NextLevel int
-}
-
-// EscalationNotifications manages escalation notifications.
-
-type EscalationNotifications struct {
-	channels map[string]NotificationChannel
-
-	templates map[string]*EscalationTemplate
-}
-
-// EscalationTemplate defines escalation notification templates.
-
-type EscalationTemplate struct {
-	ID string
-
-	Subject string
-
-	Body string
-
-	Urgency string
-
-	Channels []string
-}
-
-// IncidentCommunication manages incident communications.
-
-type IncidentCommunication struct {
-	channels map[string]*CommunicationChannel
-
-	messages []*IncidentMessage
-
-	statusPages map[string]*StatusPage
-
-	notifications *CommunicationNotifications
-}
-
-// CommunicationChannel represents a communication channel.
-
-type CommunicationChannel struct {
-	ID string
-
-	Type string // EMAIL, SLACK, SMS, WEBHOOK
-
-	Name string
-
-	Endpoint string
-
-	Enabled bool
-
-	Config map[string]interface{}
-}
-
-// IncidentMessage represents an incident communication message.
-
-type IncidentMessage struct {
-	ID string
-
-	IncidentID string
-
-	Type string // UPDATE, NOTIFICATION, RESOLUTION
-
-	Subject string
-
-	Content string
-
-	Channels []string
-
-	Recipients []string
-
-	SentAt time.Time
-
-	SentBy string
-
-	DeliveryStatus map[string]string
-}
-
-// StatusPage represents an incident status page.
-
-type StatusPage struct {
-	ID string
-
-	URL string
-
-	Title string
-
-	Description string
-
-	Incidents []string
-
-	Status string
-
-	LastUpdate time.Time
-
-	Subscribers []string
-}
-
-// CommunicationNotifications manages communication notifications.
-
-type CommunicationNotifications struct {
-	subscribers map[string]*CommunicationSubscriber
-
-	preferences map[string]*NotificationPreferences
-}
-
-// CommunicationSubscriber represents a communication subscriber.
-
-type CommunicationSubscriber struct {
-	ID string
-
-	Name string
-
-	Email string
-
-	Phone string
-
-	Preferences *NotificationPreferences
-
-	Active bool
-}
-
-// NotificationPreferences represents notification preferences.
-
-type NotificationPreferences struct {
-	Channels []string
-
-	Severity []string
-
-	Types []string
-
-	Frequency string
-
-	QuietHours *QuietHours
-}
-
-// QuietHours represents quiet hours for notifications.
-
-type QuietHours struct {
-	Enabled bool
-
-	StartTime string
-
-	EndTime string
-
-	Timezone string
-
-	Days []string
-}
-
-// DigitalForensics provides digital forensics capabilities.
-
-type DigitalForensics struct {
-	tools map[string]*ForensicTool
-
-	evidence map[string]*DigitalEvidence
-
-	chainOfCustody *ChainOfCustodyManager
-
-	analysis *ForensicAnalysis
-
-	reporting *ForensicReporting
-}
-
-// ForensicTool represents a digital forensics tool.
-
-type ForensicTool struct {
-	ID string
-
-	Name string
-
-	Type string // IMAGING, ANALYSIS, RECOVERY, TIMELINE
-
-	Version string
-
-	Description string
-
-	Capabilities []string
-
-	Licensed bool
-
-	Available bool
-}
-
-// DigitalEvidence represents digital evidence.
-
-type DigitalEvidence struct {
-	ID string
-
-	Type string // DISK_IMAGE, MEMORY_DUMP, NETWORK_CAPTURE, LOG_FILE
-
-	Description string
-
-	Source string
-
-	Size int64
-
-	Hash map[string]string // algorithm -> hash
-
-	CollectedAt time.Time
-
-	CollectedBy string
-
-	Location string
-
-	ChainOfCustody []*CustodyRecord
-
-	AnalysisResults []*AnalysisResult
-
-	Sealed bool
-}
-
-// ChainOfCustodyManager manages evidence chain of custody.
-
-type ChainOfCustodyManager struct {
-	records map[string][]*CustodyRecord
-
-	custodians map[string]*Custodian
-
-	transfers []*CustodyTransfer
-}
-
-// Custodian represents a person in the chain of custody.
-
-type Custodian struct {
-	ID string
-
-	Name string
-
-	Role string
-
-	Organization string
-
-	ContactInfo string
-
-	Authorized bool
-}
-
-// CustodyTransfer represents a custody transfer.
-
-type CustodyTransfer struct {
-	ID string
-
-	EvidenceID string
-
-	FromCustodian string
-
-	ToCustodian string
-
-	Timestamp time.Time
-
-	Reason string
-
-	Witness string
-
-	Signature string
-
-	Notes string
-}
-
-// ForensicAnalysis provides forensic analysis capabilities.
-
-type ForensicAnalysis struct {
-	analyzers map[string]*ForensicAnalyzer
-
-	workflows map[string]*AnalysisWorkflow
-
-	results map[string]*AnalysisResult
-
-	timelines map[string]*ForensicTimeline
-}
-
-// ForensicAnalyzer represents a forensic analyzer.
-
-type ForensicAnalyzer struct {
-	ID string
-
-	Name string
-
-	Type string
-
-	Capabilities []string
-
-	Enabled bool
-
-	Config map[string]interface{}
-}
-
-// AnalysisWorkflow represents a forensic analysis workflow.
-
-type AnalysisWorkflow struct {
-	ID string
-
-	Name string
-
-	Steps []*AnalysisStep
-
-	Status string
-
-	StartedAt time.Time
-
-	CompletedAt time.Time
-
-	Results []*AnalysisResult
-}
-
-// AnalysisStep represents a step in analysis workflow.
-
-type AnalysisStep struct {
-	ID string
-
-	Name string
-
-	Type string
-
-	Analyzer string
-
-	Parameters map[string]interface{}
-
-	Dependencies []string
-
-	Status string
-
-	Result *AnalysisResult
-}
-
-// AnalysisResult represents the result of forensic analysis.
-
-type AnalysisResult struct {
-	ID string
-
-	AnalyzerID string
-
-	EvidenceID string
-
-	Type string
-
-	Summary string
-
-	Findings []*ForensicFinding
-
-	Artifacts []*DigitalArtifact
-
-	Timeline *ForensicTimeline
-
-	Confidence float64
-
-	CreatedAt time.Time
-
-	CreatedBy string
-}
-
-// ForensicFinding represents a forensic finding.
-
-type ForensicFinding struct {
-	ID string
-
-	Type string
-
-	Description string
-
-	Significance string
-
-	Evidence []string
-
-	Location string
-
-	Timestamp time.Time
-
-	Confidence float64
-
-	Tags []string
-}
-
-// DigitalArtifact represents a digital artifact.
-
-type DigitalArtifact struct {
-	ID string
-
-	Type string
-
-	Name string
-
-	Path string
-
-	Size int64
-
-	Hash map[string]string
-
-	Timestamps map[string]time.Time
-
-	Metadata map[string]interface{}
-
-	Content []byte
-
-	Related []string
-}
-
-// ForensicTimeline represents a forensic timeline.
-
-type ForensicTimeline struct {
-	ID string
-
-	Name string
-
-	Events []*TimelineEvent
-
-	StartTime time.Time
-
-	EndTime time.Time
-
-	Sources []string
-
-	CreatedAt time.Time
-}
-
-// TimelineEvent represents an event in a forensic timeline.
-
-type TimelineEvent struct {
-	ID string
-
-	Timestamp time.Time
-
-	Type string
-
-	Source string
-
-	Description string
-
-	Evidence []string
-
-	Significance string
-
-	Tags []string
-}
-
-// ForensicReporting provides forensic reporting.
-
-type ForensicReporting struct {
-	templates map[string]*ForensicReportTemplate
-
-	generators map[string]*ForensicReportGenerator
-
-	reports map[string]*ForensicReport
-}
-
-// ForensicReportTemplate defines forensic report templates.
-
-type ForensicReportTemplate struct {
-	ID string
-
-	Name string
-
-	Type string
-
-	Sections []*ReportSection
-
-	Format string
-
-	Template string
-}
-
-// ReportSection represents a section in a forensic report.
-
-type ReportSection struct {
-	ID string
-
-	Title string
-
-	Content string
-
-	Type string
-
-	Required bool
-
-	Order int
-}
-
-// ForensicReportGenerator generates forensic reports.
-
-type ForensicReportGenerator interface {
-	GenerateReport(ctx context.Context, template *ForensicReportTemplate, data interface{}) (*ForensicReport, error)
-
-	GetSupportedFormats() []string
-}
-
-// ForensicReport represents a generated forensic report.
-
-type ForensicReport struct {
-	ID string
-
-	IncidentID string
-
-	Type string
-
-	Title string
-
-	Summary string
-
-	Findings []*ForensicFinding
-
-	Evidence []string
-
-	Conclusions []string
-
-	Recommendations []string
-
-	Appendices []*ReportAppendix
-
-	GeneratedAt time.Time
-
-	GeneratedBy string
-
-	Content []byte
-
-	Format string
-}
-
-// ReportAppendix represents a report appendix.
-
-type ReportAppendix struct {
-	ID string
-
-	Title string
-
-	Type string
-
-	Content []byte
-
-	Filename string
-}
-
-// IncidentResponseConfig holds incident response configuration.
-
-type IncidentResponseConfig struct {
-	DefaultPlaybook string
-
-	AutoAssignments bool
-
-	EscalationEnabled bool
-
-	CommunicationChannels []string
-
-	ForensicsEnabled bool
-
-	DocumentationRequired bool
-
-	ReviewRequired bool
-
-	MetricsEnabled bool
-}
-
-// SecurityMetrics holds Prometheus metrics for security management.
-
-type SecurityMetrics struct {
-	AuthenticationAttempts *prometheus.CounterVec
-
-	AuthenticationFailures *prometheus.CounterVec
-
-	CertificatesManaged prometheus.Gauge
-
-	CertificatesExpiring prometheus.Gauge
-
-	SecurityAlerts *prometheus.CounterVec
-
-	ThreatDetections *prometheus.CounterVec
-
-	ComplianceScore *prometheus.GaugeVec
-
-	VulnerabilitiesFound *prometheus.CounterVec
-
-	IncidentsActive prometheus.Gauge
-
-	IncidentsResolved prometheus.Counter
-}
-
-// Configuration structures and additional types would continue...
-
-// For brevity, I'll include the main interface methods.
-
-// NewComprehensiveSecurityManager creates a new security manager.
-
-func NewComprehensiveSecurityManager(config *SecurityManagerConfig) *ComprehensiveSecurityManager {
-	if config == nil {
-		config = &SecurityManagerConfig{
-			AuthenticationMethods: []string{"password", "certificate"},
-
-			SessionTimeout: 30 * time.Minute,
-
-			MaxFailedAttempts: 3,
-
-			LockoutDuration: 15 * time.Minute,
-
-			AuditLogRetention: 90 * 24 * time.Hour,
-
-			ThreatDetectionEnabled: true,
-
-			IntrusionDetectionEnabled: true,
-
-			VulnerabilityScanInterval: 24 * time.Hour,
-		}
-	}
-
-	csm := &ComprehensiveSecurityManager{
-		config: config,
-
-		securityMetrics: initializeSecurityMetrics(),
-
-		stopChan: make(chan struct{}),
-	}
-
-	// Initialize all security components.
-
-	csm.certificateManager = NewCertificateLifecycleManager(config.CertificateAuthority)
-
-	csm.authenticationMgr = NewAuthenticationManager(&AuthenticationConfig{
-		SessionTimeout: config.SessionTimeout,
-
-		MaxFailedAttempts: config.MaxFailedAttempts,
-
-		LockoutDuration: config.LockoutDuration,
-	})
-
-	csm.authorizationMgr = NewAuthorizationManager()
-
-	csm.encryptionMgr = NewEncryptionManager(config.EncryptionStandards)
-
-	csm.keyManagementService = NewKeyManagementService()
-
-	csm.securityAuditMgr = NewSecurityAuditManager(config.AuditLogRetention)
-
-	csm.securityPolicyEngine = NewDefaultSecurityPolicyEngine()
-
-	csm.secureChannelMgr = NewSecureChannelManager(&SecureChannelConfig{})
-
-	if config.IntrusionDetectionEnabled {
-		csm.intrusionDetection = NewIntrusionDetectionSystem(&IDSConfig{})
-	}
-
-	if config.ThreatDetectionEnabled {
-		csm.threatDetectionMgr = NewThreatDetectionManager(&ThreatDetectionConfig{})
-	}
-
-	csm.complianceMonitor = NewComplianceMonitor(&ComplianceConfig{
-		EnabledFrameworks: config.ComplianceModes,
-	})
-
-	csm.vulnerabilityScanner = NewVulnerabilityScanner(&VulnerabilityScanConfig{
-		ScanSchedule: map[string]string{"daily": "02:00"},
-	})
-
-	csm.incidentResponseMgr = NewIncidentResponseManager(&IncidentResponseConfig{})
-
-	return csm
-}
-
-// Start starts the security manager.
-
-func (csm *ComprehensiveSecurityManager) Start(ctx context.Context) error {
-	csm.mutex.Lock()
-
-	defer csm.mutex.Unlock()
-
-	if csm.running {
-		return fmt.Errorf("security manager already running")
-	}
-
-	logger := log.FromContext(ctx)
-
-	logger.Info("starting comprehensive security manager")
-
-	// Start certificate lifecycle management.
-
-	if err := csm.certificateManager.Start(ctx); err != nil {
-		return fmt.Errorf("failed to start certificate manager: %w", err)
-	}
-
-	// Start intrusion detection.
-
-	if csm.intrusionDetection != nil {
-		if err := csm.intrusionDetection.Start(ctx); err != nil {
-			logger.Error(err, "failed to start intrusion detection")
-		}
-	}
-
-	// Start threat detection.
-
-	if csm.threatDetectionMgr != nil {
-		if err := csm.threatDetectionMgr.Start(ctx); err != nil {
-			logger.Error(err, "failed to start threat detection")
-		}
-	}
-
-	// Start vulnerability scanning.
-
-	if err := csm.vulnerabilityScanner.Start(ctx); err != nil {
-		logger.Error(err, "failed to start vulnerability scanner")
-	}
-
-	// Start compliance monitoring.
-
-	if err := csm.complianceMonitor.Start(ctx); err != nil {
-		logger.Error(err, "failed to start compliance monitor")
-	}
-
-	csm.running = true
-
-	logger.Info("comprehensive security manager started successfully")
-
-	return nil
-}
-
-// Stop stops the security manager.
-
-func (csm *ComprehensiveSecurityManager) Stop(ctx context.Context) error {
-	csm.mutex.Lock()
-
-	defer csm.mutex.Unlock()
-
-	if !csm.running {
-		return nil
-	}
-
-	logger := log.FromContext(ctx)
-
-	logger.Info("stopping comprehensive security manager")
-
-	close(csm.stopChan)
-
-	// Stop all components.
-
-	if csm.certificateManager != nil {
-		csm.certificateManager.Stop(ctx)
-	}
-
-	if csm.intrusionDetection != nil {
-		csm.intrusionDetection.Stop(ctx)
-	}
-
-	if csm.threatDetectionMgr != nil {
-		csm.threatDetectionMgr.Stop(ctx)
-	}
-
-	if csm.vulnerabilityScanner != nil {
-		csm.vulnerabilityScanner.Stop(ctx)
-	}
-
-	if csm.complianceMonitor != nil {
-		csm.complianceMonitor.Stop(ctx)
-	}
-
-	csm.running = false
-
-	logger.Info("comprehensive security manager stopped")
-
-	return nil
-}
-
-// Core security operations.
-
-// AuthenticateUser authenticates a user.
-
-func (csm *ComprehensiveSecurityManager) AuthenticateUser(ctx context.Context, credentials *AuthCredentials) (*AuthResult, error) {
-	return csm.authenticationMgr.Authenticate(ctx, credentials)
-}
-
-// AuthorizeAccess authorizes access to a resource.
-
-func (csm *ComprehensiveSecurityManager) AuthorizeAccess(ctx context.Context, request *AccessRequest) (*AccessDecision, error) {
-	return csm.authorizationMgr.Authorize(ctx, request)
-}
-
-// IssueCertificate issues a new certificate.
-
-func (csm *ComprehensiveSecurityManager) IssueCertificate(ctx context.Context, request *CertificateRequest) (*ManagedCertificate, error) {
-	return csm.certificateManager.IssueCertificate(ctx, request)
-}
-
-// GetSecurityStatus returns overall security status.
-
-func (csm *ComprehensiveSecurityManager) GetSecurityStatus(ctx context.Context) (*SecurityStatus, error) {
-	// Implementation would aggregate status from all security components.
-
-	status := &SecurityStatus{
-		ComplianceLevel: "HIGH",
-
-		ActiveThreats: []string{},
-
-		LastAudit: time.Now().Add(-24 * time.Hour),
-
-		Metrics: json.RawMessage(`{}`),
-	}
-
-	return status, nil
-}
-
-// Helper methods and placeholder implementations.
-
-func (csm *ComprehensiveSecurityManager) getActiveAlertCount() int {
-	// Would aggregate from all security components.
-
-	return 0
-}
-
-func initializeSecurityMetrics() *SecurityMetrics {
-	return &SecurityMetrics{
-		AuthenticationAttempts: promauto.NewCounterVec(prometheus.CounterOpts{
-			Name: "oran_security_auth_attempts_total",
-
-			Help: "Total number of authentication attempts",
-		}, []string{"method", "result"}),
-
-		AuthenticationFailures: promauto.NewCounterVec(prometheus.CounterOpts{
-			Name: "oran_security_auth_failures_total",
-
-			Help: "Total number of authentication failures",
-		}, []string{"method", "reason"}),
-
-		CertificatesManaged: promauto.NewGauge(prometheus.GaugeOpts{
-			Name: "oran_security_certificates_managed",
-
-			Help: "Number of certificates under management",
-		}),
-
-		CertificatesExpiring: promauto.NewGauge(prometheus.GaugeOpts{
-			Name: "oran_security_certificates_expiring",
-
-			Help: "Number of certificates expiring soon",
-		}),
-
-		SecurityAlerts: promauto.NewCounterVec(prometheus.CounterOpts{
-			Name: "oran_security_alerts_total",
-
-			Help: "Total number of security alerts",
-		}, []string{"type", "severity"}),
-
-		ThreatDetections: promauto.NewCounterVec(prometheus.CounterOpts{
-			Name: "oran_security_threats_detected_total",
-
-			Help: "Total number of threats detected",
-		}, []string{"type", "severity"}),
-
-		ComplianceScore: promauto.NewGaugeVec(prometheus.GaugeOpts{
-			Name: "oran_security_compliance_score",
-
-			Help: "Compliance score by framework",
-		}, []string{"framework"}),
-
-		VulnerabilitiesFound: promauto.NewCounterVec(prometheus.CounterOpts{
-			Name: "oran_security_vulnerabilities_found_total",
-
-			Help: "Total number of vulnerabilities found",
-		}, []string{"severity", "type"}),
-
-		IncidentsActive: promauto.NewGauge(prometheus.GaugeOpts{
-			Name: "oran_security_incidents_active",
-
-			Help: "Number of active security incidents",
-		}),
-
-		IncidentsResolved: promauto.NewCounter(prometheus.CounterOpts{
-			Name: "oran_security_incidents_resolved_total",
-
-			Help: "Total number of resolved security incidents",
-		}),
-	}
-}
-
-// Placeholder implementations for major security components.
-
-// In production, each would be fully implemented with comprehensive functionality.
-
-// CertificateRequest represents a certificate issuance request.
-
-type CertificateRequest struct {
-	Subject pkix.Name
-
-	AlternativeNames []string
-
-	KeySize int
-
-	Usage []string
-
-	ValidityPeriod time.Duration
-
-	Template string
-}
-
-// NewCertificateLifecycleManager performs newcertificatelifecyclemanager operation.
-
-func NewCertificateLifecycleManager(caConfig *CAConfig) *CertificateLifecycleManager {
-	return &CertificateLifecycleManager{
-		certificates: make(map[string]*ManagedCertificate),
-
-		autoRenewal: &AutoRenewalService{},
-
-		validationService: &CertificateValidationService{},
-	}
-}
-
-// Start performs start operation.
-
-func (clm *CertificateLifecycleManager) Start(ctx context.Context) error { return nil }
-
-// Stop performs stop operation.
-
-func (clm *CertificateLifecycleManager) Stop(ctx context.Context) error { return nil }
-
-// IssueCertificate performs issuecertificate operation.
-
-func (clm *CertificateLifecycleManager) IssueCertificate(ctx context.Context, request *CertificateRequest) (*ManagedCertificate, error) {
-	return &ManagedCertificate{}, nil
-}
-
-// GetCertificateCount performs getcertificatecount operation.
-
-func (clm *CertificateLifecycleManager) GetCertificateCount() int { return len(clm.certificates) }
-
-// NewAuthenticationManager performs newauthenticationmanager operation.
-
-func NewAuthenticationManager(config *AuthenticationConfig) *AuthenticationManager {
-	return &AuthenticationManager{
-		methods: make(map[string]AuthenticationMethod),
-
-		sessions: make(map[string]*AuthSession),
-
-		failedAttempts: make(map[string]*FailedAttemptTracker),
-
-		config: config,
-	}
-}
-
-// Authenticate performs authenticate operation.
-
-func (am *AuthenticationManager) Authenticate(ctx context.Context, credentials *AuthCredentials) (*AuthResult, error) {
-	return &AuthResult{Success: true}, nil
-}
-
-// GetActiveSessions performs getactivesessions operation.
-
-func (am *AuthenticationManager) GetActiveSessions() int { return len(am.sessions) }
-
-// NewAuthorizationManager performs newauthorizationmanager operation.
-
-func NewAuthorizationManager() *AuthorizationManager {
-	return &AuthorizationManager{
-		rbacEngine: &RoleBasedAccessControl{},
-
-		abacEngine: &AttributeBasedAccessControl{},
-
-		policyEngine: &PolicyEngine{},
-
-		accessDecisions: make(map[string]*AccessDecision),
-	}
-}
-
-// Authorize performs authorize operation.
-
-func (azm *AuthorizationManager) Authorize(ctx context.Context, request *AccessRequest) (*AccessDecision, error) {
-	return &AccessDecision{Decision: "PERMIT"}, nil
-}
-
-// NewEncryptionManager performs newencryptionmanager operation.
-
-func NewEncryptionManager(config *EncryptionConfig) *EncryptionManager {
-	return &EncryptionManager{
-		encryptors: make(map[string]Encryptor),
-
-		config: config,
-	}
-}
-
-// NewKeyManagementService performs newkeymanagementservice operation.
-
-func NewKeyManagementService() *KeyManagementService {
-	return &KeyManagementService{
-		keys: make(map[string]*CryptographicKey),
-
-		keyPolicies: make(map[string]*KeyPolicy),
-	}
-}
-
-// NewSecurityAuditManager performs newsecurityauditmanager operation.
-
-func NewSecurityAuditManager(retention time.Duration) *SecurityAuditManager {
-	return &SecurityAuditManager{
-		auditLog: &AuditLog{entries: make([]*AuditEntry, 0)},
-
-		auditPolicies: make([]*AuditPolicy, 0),
-	}
-}
-
-// NewDefaultSecurityPolicyEngine performs newdefaultsecuritypolicyengine operation.
-
-func NewDefaultSecurityPolicyEngine() security.SecurityPolicyEngine {
-	return &DefaultSecurityPolicyEngine{
-		policies: make(map[string]*security.SecurityPolicy),
-
-		evaluators: make([]PolicyEvaluator, 0),
-
-		cache: &PolicyCache{decisions: make(map[string]*CachedDecision)},
-	}
-}
-
-// DefaultSecurityPolicyEngine provides a default implementation of SecurityPolicyEngine.
-
-type DefaultSecurityPolicyEngine struct {
-	policies map[string]*security.SecurityPolicy
-
-	evaluators []PolicyEvaluator
-
-	cache *PolicyCache
-
-	mutex sync.RWMutex
-}
-
-// EvaluatePolicy performs evaluatepolicy operation.
-
-func (dpe *DefaultSecurityPolicyEngine) EvaluatePolicy(request *security.PolicyRequest) (*security.PolicyDecision, error) {
-	// Simplified implementation for compilation.
-
-	return &security.PolicyDecision{
-		Decision: "PERMIT",
-
-		Reason: "Default allow policy",
-	}, nil
-}
-
-// AddPolicy performs addpolicy operation.
-
-func (dpe *DefaultSecurityPolicyEngine) AddPolicy(policy *security.SecurityPolicy) error {
-	dpe.mutex.Lock()
-
-	defer dpe.mutex.Unlock()
-
-	dpe.policies[policy.ID] = policy
-
-	return nil
-}
-
-// RemovePolicy performs removepolicy operation.
-
-func (dpe *DefaultSecurityPolicyEngine) RemovePolicy(policyID string) error {
-	dpe.mutex.Lock()
-
-	defer dpe.mutex.Unlock()
-
-	delete(dpe.policies, policyID)
-
-	return nil
-}
-
-// UpdatePolicy performs updatepolicy operation.
-
-func (dpe *DefaultSecurityPolicyEngine) UpdatePolicy(policy *security.SecurityPolicy) error {
-	dpe.mutex.Lock()
-
-	defer dpe.mutex.Unlock()
-
-	dpe.policies[policy.ID] = policy
-
-	return nil
-}
-
-// GetPolicy performs getpolicy operation.
-
-func (dpe *DefaultSecurityPolicyEngine) GetPolicy(policyID string) (*security.SecurityPolicy, error) {
-	dpe.mutex.RLock()
-
-	defer dpe.mutex.RUnlock()
-
-	if policy, exists := dpe.policies[policyID]; exists {
-		return policy, nil
-	}
-
-	return nil, fmt.Errorf("policy not found: %s", policyID)
-}
-
-// ListPolicies performs listpolicies operation.
-
-func (dpe *DefaultSecurityPolicyEngine) ListPolicies(filter *security.PolicyFilter) ([]*security.SecurityPolicy, error) {
-	dpe.mutex.RLock()
-
-	defer dpe.mutex.RUnlock()
-
-	var result []*security.SecurityPolicy
-
-	for _, policy := range dpe.policies {
-		result = append(result, policy)
-	}
-
-	return result, nil
-}
-
-// NewSecureChannelManager performs newsecurechannelmanager operation.
-
-func NewSecureChannelManager(config *SecureChannelConfig) *SecureChannelManager {
-	return &SecureChannelManager{
-		channels: make(map[string]*SecureChannel),
-
-		config: config,
-	}
-}
-
-// NewIntrusionDetectionSystem performs newintrusiondetectionsystem operation.
-
-func NewIntrusionDetectionSystem(config *IDSConfig) *IntrusionDetectionSystem {
-	return &IntrusionDetectionSystem{
-		sensors: make(map[string]*IDSSensor),
-
-		rules: make([]*IDSRule, 0),
-
-		alerts: make(chan *SecurityAlert, 1000),
-
-		config: config,
-	}
-}
-
-// Start performs start operation.
-
-func (ids *IntrusionDetectionSystem) Start(ctx context.Context) error { ids.running = true; return nil }
-
-// Stop performs stop operation.
-
-func (ids *IntrusionDetectionSystem) Stop(ctx context.Context) error { ids.running = false; return nil }
-
-// NewThreatDetectionManager performs newthreatdetectionmanager operation.
-
-func NewThreatDetectionManager(config *ThreatDetectionConfig) *ThreatDetectionManager {
-	return &ThreatDetectionManager{
-		detectors: make(map[string]*ThreatDetector),
-
-		threatDatabase: &ThreatDatabase{},
-
-		config: config,
-	}
-}
-
-// Start performs start operation.
-
-func (tdm *ThreatDetectionManager) Start(ctx context.Context) error { tdm.running = true; return nil }
-
-// Stop performs stop operation.
-
-func (tdm *ThreatDetectionManager) Stop(ctx context.Context) error { tdm.running = false; return nil }
-
-// NewComplianceMonitor performs newcompliancemonitor operation.
-
-func NewComplianceMonitor(config *ComplianceConfig) *ComplianceMonitor {
-	return &ComplianceMonitor{
-		frameworks: make(map[string]*ComplianceFramework),
-
-		assessments: make(map[string]*ComplianceAssessment),
-
-		controls: make(map[string]*ComplianceControl),
-
-		config: config,
-	}
-}
-
-// Start performs start operation.
-
-func (cm *ComplianceMonitor) Start(ctx context.Context) error { return nil }
-
-// Stop performs stop operation.
-
-func (cm *ComplianceMonitor) Stop(ctx context.Context) error { return nil }
-
-// GetOverallScore performs getoverallscore operation.
-
-func (cm *ComplianceMonitor) GetOverallScore() float64 { return 85.5 }
-
-// NewVulnerabilityScanner performs newvulnerabilityscanner operation.
-
-func NewVulnerabilityScanner(config *VulnerabilityScanConfig) *VulnerabilityScanner {
-	return &VulnerabilityScanner{
-		scanners: make(map[string]*VulnScanner),
-
-		database: &VulnerabilityDatabase{},
-
-		config: config,
-	}
-}
-
-// Start performs start operation.
-
-func (vs *VulnerabilityScanner) Start(ctx context.Context) error { vs.running = true; return nil }
-
-// Stop performs stop operation.
-
-func (vs *VulnerabilityScanner) Stop(ctx context.Context) error { vs.running = false; return nil }
-
-// GetVulnerabilityCount performs getvulnerabilitycount operation.
-
-func (vs *VulnerabilityScanner) GetVulnerabilityCount() int { return 0 }
-
-// NewIncidentResponseManager performs newincidentresponsemanager operation.
-
-func NewIncidentResponseManager(config *IncidentResponseConfig) *IncidentResponseManager {
-	return &IncidentResponseManager{
-		incidents: make(map[string]*SecurityIncident),
-
-		playbooks: make(map[string]*IncidentPlaybook),
-
-		responders: make(map[string]*IncidentResponder),
-
-		workflows: make(map[string]*ResponseWorkflow),
-
-		config: config,
-	}
-}
-
-// Additional configuration types.
-
-type SecureChannelConfig struct {
-	DefaultTLSVersion uint16
-
-	CipherSuites []uint16
-
-	MaxConnections int
-
-	IdleTimeout time.Duration
-}
-
