+++ pkg/oran/o1/security_manager_extended.go
@@ -1,0 +1,1317 @@
+// Package o1 - Extended security types (part 2)
+// This file is part of the security_manager.go split.
+package o1
+
+import (
+	_ "crypto/tls" // imported for side effects
+	_ "context"
+	"crypto/tls"
+	_ "crypto/x509"
+	_ "crypto/x509/pkix"
+	_ "encoding/json"
+	_ "fmt"
+	_ "math/big"
+	"net"
+	"sync"
+	"time"
+
+	_ "github.com/prometheus/client_golang/prometheus"
+	_ "github.com/prometheus/client_golang/prometheus/promauto"
+	_ "sigs.k8s.io/controller-runtime/pkg/log"
+
+	_ "github.com/thc1006/nephoran-intent-operator/pkg/oran"
+	"github.com/thc1006/nephoran-intent-operator/pkg/oran/o1/security"
+)
+
+
+// ActiveMitigation represents an active threat mitigation.
+
+type ActiveMitigation struct {
+	ID string
+
+	ThreatID string
+
+	StrategyID string
+
+	Status string
+
+	StartedAt time.Time
+
+	CompletedAt time.Time
+
+	Actions []*ActionExecution
+
+	Results map[string]interface{}
+}
+
+// MitigationPlaybook defines automated mitigation procedures.
+
+type MitigationPlaybook struct {
+	ID string
+
+	Name string
+
+	ThreatTypes []string
+
+	TriggerConditions []string
+
+	Strategies []string
+
+	Escalation []*EscalationRule
+
+	Approval *ApprovalRequirement
+
+	Enabled bool
+}
+
+// ApprovalRequirement defines approval requirements.
+
+type ApprovalRequirement struct {
+	Required bool
+
+	Approvers []string
+
+	Threshold int
+
+	Timeout time.Duration
+
+	EscalationLevel int
+}
+
+// ThreatDetectionConfig holds threat detection configuration.
+
+type ThreatDetectionConfig struct {
+	EnabledDetectors []string
+
+	UpdateInterval time.Duration
+
+	ConfidenceThreshold float64
+
+	AutoMitigation bool
+
+	NotificationChannels []string
+}
+
+// RiskAssessmentConfig holds risk assessment configuration.
+
+type RiskAssessmentConfig struct {
+	DefaultModel string
+
+	AssessmentInterval time.Duration
+
+	AutoAssessment bool
+
+	RiskThresholds map[string]float64
+
+	ReportingInterval time.Duration
+}
+
+// ComplianceMonitor monitors security compliance.
+
+type ComplianceMonitor struct {
+	frameworks map[string]*ComplianceFramework
+
+	assessments map[string]*ComplianceAssessment
+
+	controls map[string]*ComplianceControl
+
+	scanner *ComplianceScanner
+
+	reporter *ComplianceReporter
+
+	config *ComplianceConfig
+
+	mutex sync.RWMutex
+}
+
+// ComplianceFramework represents a compliance framework.
+
+type ComplianceFramework struct {
+	ID string
+
+	Name string
+
+	Version string
+
+	Description string
+
+	Authority string
+
+	Domains []*ComplianceDomain
+
+	Controls []string
+
+	Applicable bool
+
+	Mandatory bool
+
+	CreatedAt time.Time
+
+	UpdatedAt time.Time
+}
+
+// ComplianceDomain represents a domain within a framework.
+
+type ComplianceDomain struct {
+	ID string
+
+	Name string
+
+	Description string
+
+	Controls []string
+
+	Weight float64
+}
+
+// ComplianceControl represents a compliance control.
+
+type ComplianceControl struct {
+	ID string
+
+	Framework string
+
+	Number string
+
+	Title string
+
+	Description string
+
+	Type string // PREVENTIVE, DETECTIVE, CORRECTIVE
+
+	Category string // TECHNICAL, ADMINISTRATIVE, PHYSICAL
+
+	Implementation string
+
+	Evidence []string
+
+	TestProcedures []string
+
+	Frequency string
+
+	Owner string
+
+	Status string // NOT_IMPLEMENTED, PARTIALLY_IMPLEMENTED, IMPLEMENTED, NOT_APPLICABLE
+
+	LastAssessed time.Time
+
+	NextAssessment time.Time
+}
+
+// ComplianceAssessment represents a compliance assessment.
+
+type ComplianceAssessment struct {
+	ID string
+
+	Framework string
+
+	AssessmentType string // SELF, THIRD_PARTY, REGULATORY
+
+	Assessor string
+
+	StartDate time.Time
+
+	EndDate time.Time
+
+	Status string // PLANNED, IN_PROGRESS, COMPLETED, REMEDIATION
+
+	OverallScore float64
+
+	ControlResults map[string]*ControlResult
+
+	Findings []*ComplianceFinding
+
+	Remediation []*RemediationItem
+
+	Evidence []*ComplianceEvidence
+
+	Report *ComplianceReport
+}
+
+// ControlResult represents the result of a control assessment.
+
+type ControlResult struct {
+	ControlID string
+
+	Status string // COMPLIANT, NON_COMPLIANT, PARTIALLY_COMPLIANT, NOT_APPLICABLE
+
+	Score float64
+
+	Evidence []string
+
+	Findings []string
+
+	Recommendations []string
+
+	TestResults []*TestResult
+}
+
+// TestResult represents the result of a control test.
+
+type TestResult struct {
+	TestID string
+
+	Status string
+
+	Score float64
+
+	Evidence []string
+
+	Notes string
+
+	TestedAt time.Time
+
+	TestedBy string
+}
+
+// ComplianceFinding represents a compliance finding.
+
+type ComplianceFinding struct {
+	ID string
+
+	Type string // DEFICIENCY, WEAKNESS, VIOLATION
+
+	Severity string // LOW, MEDIUM, HIGH, CRITICAL
+
+	Control string
+
+	Description string
+
+	Impact string
+
+	Likelihood string
+
+	Risk string
+
+	Evidence []string
+
+	Recommendations []string
+
+	Status string // OPEN, IN_REMEDIATION, CLOSED
+
+	IdentifiedAt time.Time
+
+	DueDate time.Time
+
+	Owner string
+}
+
+// RemediationItem represents a remediation action.
+
+type RemediationItem struct {
+	ID string
+
+	FindingID string
+
+	Action string
+
+	Description string
+
+	Priority string
+
+	Owner string
+
+	DueDate time.Time
+
+	Status string // PLANNED, IN_PROGRESS, COMPLETED, VERIFIED
+
+	EstimatedCost float64
+
+	ActualCost float64
+
+	Evidence []string
+}
+
+// ComplianceEvidence represents compliance evidence.
+
+type ComplianceEvidence struct {
+	ID string
+
+	Type string // DOCUMENT, SCREENSHOT, LOG, CONFIGURATION
+
+	Description string
+
+	Source string
+
+	CollectedAt time.Time
+
+	Hash string
+
+	Location string
+
+	Metadata map[string]interface{}
+}
+
+// ComplianceReport represents a compliance report.
+
+type ComplianceReport struct {
+	ID string
+
+	AssessmentID string
+
+	ReportType string
+
+	Framework string
+
+	GeneratedAt time.Time
+
+	ReportPeriod *TimeRange
+
+	ExecutiveSummary string
+
+	OverallScore float64
+
+	ComplianceLevel string
+
+	KeyFindings []string
+
+	Recommendations []string
+
+	Trends map[string]interface{}
+
+	Content []byte
+
+	Format string // PDF, HTML, JSON
+}
+
+// ComplianceScanner performs automated compliance scanning.
+
+type ComplianceScanner struct {
+	scanners map[string]*Scanner
+
+	schedules map[string]*ScanSchedule
+
+	results map[string]*ScanResult
+
+	config *ScannerConfig
+
+	running bool
+
+	mutex sync.RWMutex
+}
+
+// Scanner represents a compliance scanner.
+
+type Scanner struct {
+	ID string
+
+	Type string // CONFIGURATION, VULNERABILITY, POLICY
+
+	Name string
+
+	Description string
+
+	Rules []*ScanRule
+
+	Schedule string
+
+	Enabled bool
+
+	LastRun time.Time
+
+	NextRun time.Time
+}
+
+// ScanRule represents a scanning rule.
+
+type ScanRule struct {
+	ID string
+
+	Name string
+
+	Description string
+
+	Pattern string
+
+	Expected interface{}
+
+	Severity string
+
+	Control string
+
+	Enabled bool
+}
+
+// ScanSchedule represents a scan schedule.
+
+type ScanSchedule struct {
+	ScannerID string
+
+	Frequency string // DAILY, WEEKLY, MONTHLY
+
+	Time string
+
+	Timezone string
+
+	Enabled bool
+
+	LastRun time.Time
+
+	NextRun time.Time
+
+	RunCount int64
+}
+
+// ScanResult represents scan results.
+
+type ScanResult struct {
+	ID string
+
+	ScannerID string
+
+	StartTime time.Time
+
+	EndTime time.Time
+
+	Status string // RUNNING, COMPLETED, FAILED
+
+	TotalRules int
+
+	PassedRules int
+
+	FailedRules int
+
+	Score float64
+
+	Findings []*ScanFinding
+
+	Error string
+}
+
+// ScanFinding represents a scan finding.
+
+type ScanFinding struct {
+	RuleID string
+
+	Status string // PASS, FAIL, ERROR, SKIP
+
+	Message string
+
+	Evidence interface{}
+
+	Severity string
+
+	Remediation string
+}
+
+// ComplianceReporter generates compliance reports.
+
+type ComplianceReporter struct {
+	templates map[string]*ReportTemplate
+
+	generators map[string]*ReportGenerator
+
+	schedulers map[string]*ReportScheduler
+
+	distributors map[string]*ReportDistributor
+}
+
+// ComplianceConfig holds compliance monitoring configuration.
+
+type ComplianceConfig struct {
+	EnabledFrameworks []string
+
+	AssessmentSchedule map[string]string
+
+	ScanningEnabled bool
+
+	ReportingEnabled bool
+
+	NotificationChannels []string
+
+	EvidenceRetention time.Duration
+}
+
+// ScannerConfig holds scanner configuration.
+
+type ScannerConfig struct {
+	MaxConcurrentScans int
+
+	ScanTimeout time.Duration
+
+	RetryAttempts int
+
+	ResultRetention time.Duration
+}
+
+// SecureChannelManager manages secure communications.
+
+type SecureChannelManager struct {
+	channels map[string]*SecureChannel
+
+	tlsConfig *tls.Config
+
+	vpnManager *VPNManager
+
+	tunnelMgr *TunnelManager
+
+	config *SecureChannelConfig
+
+	mutex sync.RWMutex
+}
+
+// SecureChannel represents a secure communication channel.
+
+type SecureChannel struct {
+	ID string
+
+	Type string // TLS, VPN, SSH, IPSEC
+
+	LocalEndpoint string
+
+	RemoteEndpoint string
+
+	Status string // ACTIVE, INACTIVE, ERROR
+
+	Encryption string
+
+	KeyExchange string
+
+	Authentication string
+
+	CreatedAt time.Time
+
+	LastActivity time.Time
+
+	BytesSent uint64
+
+	BytesReceived uint64
+}
+
+// VPNManager manages VPN connections.
+
+type VPNManager struct {
+	connections map[string]*VPNConnection
+
+	profiles map[string]*VPNProfile
+
+	config *security.VPNConfig
+
+	mutex sync.RWMutex
+}
+
+// VPNConnection represents a VPN connection.
+
+type VPNConnection struct {
+	ID string
+
+	ProfileID string
+
+	Status string
+
+	LocalIP net.IP
+
+	RemoteIP net.IP
+
+	Tunnel string
+
+	Encryption string
+
+	ConnectedAt time.Time
+
+	LastActivity time.Time
+
+	BytesTransferred uint64
+}
+
+// VPNProfile represents a VPN profile.
+
+type VPNProfile struct {
+	ID string
+
+	Name string
+
+	Type string // OPENVPN, IPSEC, WIREGUARD
+
+	ServerAddress string
+
+	Port int
+
+	Protocol string
+
+	Encryption string
+
+	Authentication string
+
+	Routes []string
+
+	DNS []string
+
+	Credentials *VPNCredentials
+
+	Enabled bool
+}
+
+// VPNCredentials represents VPN credentials.
+
+type VPNCredentials struct {
+	Username string
+
+	Password string
+
+	Certificate string
+
+	PrivateKey string
+
+	PSK string
+}
+
+// TunnelManager manages secure tunnels.
+
+type TunnelManager struct {
+	tunnels map[string]*SecureTunnel
+
+	config *security.TunnelConfig
+
+	mutex sync.RWMutex
+}
+
+// SecureTunnel represents a secure tunnel.
+
+type SecureTunnel struct {
+	ID string
+
+	Type string // SSH, SSL, STUNNEL
+
+	LocalPort int
+
+	RemoteHost string
+
+	RemotePort int
+
+	Status string
+
+	CreatedAt time.Time
+
+	LastActivity time.Time
+
+	Connections int64
+}
+
+// VulnerabilityScanner scans for security vulnerabilities.
+
+type VulnerabilityScanner struct {
+	scanners map[string]*VulnScanner
+
+	database *VulnerabilityDatabase
+
+	scheduler security.ScanScheduler
+
+	reporter *VulnerabilityReporter
+
+	config *VulnerabilityScanConfig
+
+	running bool
+
+	mutex sync.RWMutex
+}
+
+// VulnScanner represents a vulnerability scanner.
+
+type VulnScanner struct {
+	ID string
+
+	Type string // NETWORK, WEB, DATABASE, CONFIG
+
+	Name string
+
+	Version string
+
+	Enabled bool
+
+	LastUpdate time.Time
+
+	ScanProfiles map[string]*ScanProfile
+}
+
+// ScanProfile represents a scan profile.
+
+type ScanProfile struct {
+	ID string
+
+	Name string
+
+	Type string
+
+	Targets []string
+
+	Excludes []string
+
+	Intensity string // LOW, MEDIUM, HIGH, AGGRESSIVE
+
+	Schedule string
+
+	Notifications []string
+
+	Enabled bool
+}
+
+// VulnerabilityDatabase manages vulnerability information.
+
+type VulnerabilityDatabase struct {
+	vulnerabilities map[string]*Vulnerability
+
+	patches map[string]*SecurityPatch
+
+	advisories map[string]*SecurityAdvisory
+
+	cveDatabase *CVEDatabase
+
+	lastUpdate time.Time
+
+	mutex sync.RWMutex
+}
+
+// Vulnerability represents a security vulnerability.
+
+type Vulnerability struct {
+	ID string
+
+	CVE string
+
+	Title string
+
+	Description string
+
+	Severity string
+
+	CVSSScore float64
+
+	CVSSVector string
+
+	AffectedSystems []string
+
+	Exploitability string
+
+	Impact string
+
+	Published time.Time
+
+	Modified time.Time
+
+	References []string
+
+	Solutions []string
+
+	Workarounds []string
+
+	Status string // NEW, ASSIGNED, MODIFIED, ANALYZED, CLOSED
+}
+
+// SecurityPatch represents a security patch.
+
+type SecurityPatch struct {
+	ID string
+
+	VulnerabilityID string
+
+	Title string
+
+	Description string
+
+	Vendor string
+
+	Product string
+
+	Version string
+
+	PatchLevel string
+
+	ReleaseDate time.Time
+
+	InstallDate time.Time
+
+	Status string // AVAILABLE, INSTALLED, SUPERSEDED, WITHDRAWN
+
+	DownloadURL string
+
+	Checksum string
+
+	Prerequisites []string
+
+	RestartRequired bool
+}
+
+// SecurityAdvisory represents a security advisory.
+
+type SecurityAdvisory struct {
+	ID string
+
+	Title string
+
+	Description string
+
+	Severity string
+
+	Vendor string
+
+	Products []string
+
+	Vulnerabilities []string
+
+	Patches []string
+
+	Workarounds []string
+
+	Published time.Time
+
+	References []string
+
+	Status string
+}
+
+// CVEDatabase manages CVE information.
+
+type CVEDatabase struct {
+	entries map[string]*CVEEntry
+
+	lastUpdate time.Time
+
+	updateURL string
+
+	mutex sync.RWMutex
+}
+
+// CVEEntry represents a CVE database entry.
+
+type CVEEntry struct {
+	ID string
+
+	Description string
+
+	Published time.Time
+
+	Modified time.Time
+
+	CVSSv2 *CVSSv2Score
+
+	CVSSv3 *CVSSv3Score
+
+	References []string
+
+	CPE []string
+}
+
+// CVSSv2Score represents CVSS v2 scoring.
+
+type CVSSv2Score struct {
+	BaseScore float64
+
+	TemporalScore float64
+
+	EnvironmentalScore float64
+
+	AccessVector string
+
+	AccessComplexity string
+
+	Authentication string
+
+	ConfidentialityImpact string
+
+	IntegrityImpact string
+
+	AvailabilityImpact string
+}
+
+// CVSSv3Score represents CVSS v3 scoring.
+
+type CVSSv3Score struct {
+	BaseScore float64
+
+	TemporalScore float64
+
+	EnvironmentalScore float64
+
+	AttackVector string
+
+	AttackComplexity string
+
+	PrivilegesRequired string
+
+	UserInteraction string
+
+	Scope string
+
+	ConfidentialityImpact string
+
+	IntegrityImpact string
+
+	AvailabilityImpact string
+}
+
+// VulnerabilityReporter generates vulnerability reports.
+
+type VulnerabilityReporter struct {
+	templates map[string]*security.VulnReportTemplate
+
+	generators map[string]security.VulnReportGenerator
+
+	distributors map[string]security.VulnReportDistributor
+}
+
+// VulnerabilityScanConfig holds vulnerability scanning configuration.
+
+type VulnerabilityScanConfig struct {
+	EnabledScanners []string
+
+	ScanSchedule map[string]string
+
+	DatabaseUpdateInterval time.Duration
+
+	ReportingEnabled bool
+
+	NotificationChannels []string
+
+	MaxConcurrentScans int
+
+	ScanTimeout time.Duration
+}
+
+// IncidentResponseManager manages security incident response.
+
+type IncidentResponseManager struct {
+	incidents map[string]*SecurityIncident
+
+	playbooks map[string]*IncidentPlaybook
+
+	responders map[string]*IncidentResponder
+
+	workflows map[string]*ResponseWorkflow
+
+	escalation *IncidentEscalation
+
+	communication *IncidentCommunication
+
+	forensics *DigitalForensics
+
+	config *IncidentResponseConfig
+
+	mutex sync.RWMutex
+}
+
+// IncidentPlaybook defines incident response procedures.
+
+type IncidentPlaybook struct {
+	ID string
+
+	Name string
+
+	Description string
+
+	IncidentTypes []string
+
+	Severity string
+
+	Phases []*ResponsePhase
+
+	Roles []*IncidentRole
+
+	Tools []string
+
+	Checklists []*security.ResponseChecklist
+
+	Version string
+
+	Approved bool
+
+	CreatedAt time.Time
+
+	UpdatedAt time.Time
+}
+
+// ResponsePhase represents a phase in incident response.
+
+type ResponsePhase struct {
+	ID string
+
+	Name string
+
+	Description string
+
+	Objectives []string
+
+	Activities []*ResponseActivity
+
+	Duration time.Duration
+
+	Prerequisites []string
+
+	Deliverables []string
+
+	ExitCriteria []string
+}
+
+// ResponseActivity represents an activity in incident response.
+
+type ResponseActivity struct {
+	ID string
+
+	Name string
+
+	Description string
+
+	Type string // MANUAL, AUTOMATED, DECISION
+
+	Assignee string
+
+	Tools []string
+
+	Duration time.Duration
+
+	Dependencies []string
+
+	Outputs []string
+
+	Checklist []string
+}
+
+// IncidentRole represents a role in incident response.
+
+type IncidentRole struct {
+	ID string
+
+	Name string
+
+	Description string
+
+	Responsibilities []string
+
+	Skills []string
+
+	Authority string
+
+	ContactInfo string
+
+	Backup string
+}
+
+// IncidentResponder represents a person involved in incident response.
+
+type IncidentResponder struct {
+	ID string
+
+	Name string
+
+	Email string
+
+	Phone string
+
+	Roles []string
+
+	Skills []string
+
+	Availability string
+
+	OnCall bool
+
+	LastActive time.Time
+}
+
+// ResponseWorkflow represents an incident response workflow.
+
+type ResponseWorkflow struct {
+	ID string
+
+	IncidentID string
+
+	PlaybookID string
+
+	Status string // INITIATED, IN_PROGRESS, COMPLETED, ABORTED
+
+	CurrentPhase string
+
+	Phases map[string]*WorkflowPhase
+
+	Assignments map[string]string // activity -> assignee
+
+	StartedAt time.Time
+
+	CompletedAt time.Time
+
+	Escalated bool
+
+	Notes []string
+}
+
+// WorkflowPhase represents a workflow phase execution.
+
+type WorkflowPhase struct {
+	ID string
+
+	Status string
+
+	StartedAt time.Time
+
+	CompletedAt time.Time
+
+	Activities map[string]*ActivityExecution
+
+	Notes []string
+}
+
+// ActivityExecution represents activity execution status.
+
+type ActivityExecution struct {
+	ID string
+
+	Status string
+
+	Assignee string
+
+	StartedAt time.Time
+
+	CompletedAt time.Time
+
+	Result interface{}
+
+	Evidence []string
+
+	Notes string
+}
+
+// IncidentEscalation manages incident escalation.
+
+type IncidentEscalation struct {
+	rules []*EscalationRule
+
+	matrix *EscalationMatrix
+
+	notifications *EscalationNotifications
+}
+
+// SecurityEscalationRule defines security escalation conditions.
+
+type SecurityEscalationRule struct {
+	ID string
+
+	Name string
+
+	Conditions []string
+
+	Severity string
+
+	TimeThreshold time.Duration
+
+	Target string
+
+	Action string
+
+	Enabled bool
+}
+
+// EscalationMatrix defines escalation paths.
+
+type EscalationMatrix struct {
+	Levels map[int]*EscalationLevel
+}
+
+// EscalationLevel represents an escalation level.
+
+type EscalationLevel struct {
+	Level int
+
+	Name string
+
+	Contacts []string
+
+	TimeLimit time.Duration
+
+	Authority []string
+
+	NextLevel int
+}
+
+// EscalationNotifications manages escalation notifications.
+
+type EscalationNotifications struct {
+	channels map[string]NotificationChannel
+
+	templates map[string]*EscalationTemplate
+}
+
+// EscalationTemplate defines escalation notification templates.
+
+type EscalationTemplate struct {
+	ID string
+
+	Subject string
+
+	Body string
+
+	Urgency string
+
+	Channels []string
+}
+
+// IncidentCommunication manages incident communications.
+
+type IncidentCommunication struct {
+	channels map[string]*CommunicationChannel
+
+	messages []*IncidentMessage
+
+	statusPages map[string]*StatusPage
+
+	notifications *CommunicationNotifications
+}
+
+// CommunicationChannel represents a communication channel.
+
+type CommunicationChannel struct {
+	ID string
+
+	Type string // EMAIL, SLACK, SMS, WEBHOOK
+
+	Name string
+
+	Endpoint string
+
+	Enabled bool
+
+	Config map[string]interface{}
+}
+
+// IncidentMessage represents an incident communication message.
+
+type IncidentMessage struct {
+	ID string
+
+	IncidentID string
+
+	Type string // UPDATE, NOTIFICATION, RESOLUTION
+
+	Subject string
+
+	Content string
+
+	Channels []string
+
+	Recipients []string
+
+	SentAt time.Time
+
+	SentBy string
+
+	DeliveryStatus map[string]string
+}
+
+// StatusPage represents an incident status page.
+
+type StatusPage struct {
+	ID string
+
+	URL string
+
+	Title string
+
+	Description string
+
+	Incidents []string
+
+	Status string
+
+	LastUpdate time.Time
+
+	Subscribers []string
+}
