+++ pkg/oran/o1/security_manager_extended2.go
@@ -1,0 +1,1152 @@
+// Package o1 - Extended security types (part 3)
+// This file is part of the security_manager.go split.
+package o1
+
+import (
+	_ "crypto/tls" // imported for side effects
+	"context"
+	_ "crypto/tls"
+	_ "crypto/x509"
+	"crypto/x509/pkix"
+	"encoding/json"
+	"fmt"
+	_ "math/big"
+	_ "net"
+	"sync"
+	"time"
+
+	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/promauto"
+	"sigs.k8s.io/controller-runtime/pkg/log"
+
+	_ "github.com/thc1006/nephoran-intent-operator/pkg/oran"
+	"github.com/thc1006/nephoran-intent-operator/pkg/oran/o1/security"
+)
+
+
+// CommunicationNotifications manages communication notifications.
+
+type CommunicationNotifications struct {
+	subscribers map[string]*CommunicationSubscriber
+
+	preferences map[string]*NotificationPreferences
+}
+
+// CommunicationSubscriber represents a communication subscriber.
+
+type CommunicationSubscriber struct {
+	ID string
+
+	Name string
+
+	Email string
+
+	Phone string
+
+	Preferences *NotificationPreferences
+
+	Active bool
+}
+
+// NotificationPreferences represents notification preferences.
+
+type NotificationPreferences struct {
+	Channels []string
+
+	Severity []string
+
+	Types []string
+
+	Frequency string
+
+	QuietHours *QuietHours
+}
+
+// QuietHours represents quiet hours for notifications.
+
+type QuietHours struct {
+	Enabled bool
+
+	StartTime string
+
+	EndTime string
+
+	Timezone string
+
+	Days []string
+}
+
+// DigitalForensics provides digital forensics capabilities.
+
+type DigitalForensics struct {
+	tools map[string]*ForensicTool
+
+	evidence map[string]*DigitalEvidence
+
+	chainOfCustody *ChainOfCustodyManager
+
+	analysis *ForensicAnalysis
+
+	reporting *ForensicReporting
+}
+
+// ForensicTool represents a digital forensics tool.
+
+type ForensicTool struct {
+	ID string
+
+	Name string
+
+	Type string // IMAGING, ANALYSIS, RECOVERY, TIMELINE
+
+	Version string
+
+	Description string
+
+	Capabilities []string
+
+	Licensed bool
+
+	Available bool
+}
+
+// DigitalEvidence represents digital evidence.
+
+type DigitalEvidence struct {
+	ID string
+
+	Type string // DISK_IMAGE, MEMORY_DUMP, NETWORK_CAPTURE, LOG_FILE
+
+	Description string
+
+	Source string
+
+	Size int64
+
+	Hash map[string]string // algorithm -> hash
+
+	CollectedAt time.Time
+
+	CollectedBy string
+
+	Location string
+
+	ChainOfCustody []*CustodyRecord
+
+	AnalysisResults []*AnalysisResult
+
+	Sealed bool
+}
+
+// ChainOfCustodyManager manages evidence chain of custody.
+
+type ChainOfCustodyManager struct {
+	records map[string][]*CustodyRecord
+
+	custodians map[string]*Custodian
+
+	transfers []*CustodyTransfer
+}
+
+// Custodian represents a person in the chain of custody.
+
+type Custodian struct {
+	ID string
+
+	Name string
+
+	Role string
+
+	Organization string
+
+	ContactInfo string
+
+	Authorized bool
+}
+
+// CustodyTransfer represents a custody transfer.
+
+type CustodyTransfer struct {
+	ID string
+
+	EvidenceID string
+
+	FromCustodian string
+
+	ToCustodian string
+
+	Timestamp time.Time
+
+	Reason string
+
+	Witness string
+
+	Signature string
+
+	Notes string
+}
+
+// ForensicAnalysis provides forensic analysis capabilities.
+
+type ForensicAnalysis struct {
+	analyzers map[string]*ForensicAnalyzer
+
+	workflows map[string]*AnalysisWorkflow
+
+	results map[string]*AnalysisResult
+
+	timelines map[string]*ForensicTimeline
+}
+
+// ForensicAnalyzer represents a forensic analyzer.
+
+type ForensicAnalyzer struct {
+	ID string
+
+	Name string
+
+	Type string
+
+	Capabilities []string
+
+	Enabled bool
+
+	Config map[string]interface{}
+}
+
+// AnalysisWorkflow represents a forensic analysis workflow.
+
+type AnalysisWorkflow struct {
+	ID string
+
+	Name string
+
+	Steps []*AnalysisStep
+
+	Status string
+
+	StartedAt time.Time
+
+	CompletedAt time.Time
+
+	Results []*AnalysisResult
+}
+
+// AnalysisStep represents a step in analysis workflow.
+
+type AnalysisStep struct {
+	ID string
+
+	Name string
+
+	Type string
+
+	Analyzer string
+
+	Parameters map[string]interface{}
+
+	Dependencies []string
+
+	Status string
+
+	Result *AnalysisResult
+}
+
+// AnalysisResult represents the result of forensic analysis.
+
+type AnalysisResult struct {
+	ID string
+
+	AnalyzerID string
+
+	EvidenceID string
+
+	Type string
+
+	Summary string
+
+	Findings []*ForensicFinding
+
+	Artifacts []*DigitalArtifact
+
+	Timeline *ForensicTimeline
+
+	Confidence float64
+
+	CreatedAt time.Time
+
+	CreatedBy string
+}
+
+// ForensicFinding represents a forensic finding.
+
+type ForensicFinding struct {
+	ID string
+
+	Type string
+
+	Description string
+
+	Significance string
+
+	Evidence []string
+
+	Location string
+
+	Timestamp time.Time
+
+	Confidence float64
+
+	Tags []string
+}
+
+// DigitalArtifact represents a digital artifact.
+
+type DigitalArtifact struct {
+	ID string
+
+	Type string
+
+	Name string
+
+	Path string
+
+	Size int64
+
+	Hash map[string]string
+
+	Timestamps map[string]time.Time
+
+	Metadata map[string]interface{}
+
+	Content []byte
+
+	Related []string
+}
+
+// ForensicTimeline represents a forensic timeline.
+
+type ForensicTimeline struct {
+	ID string
+
+	Name string
+
+	Events []*TimelineEvent
+
+	StartTime time.Time
+
+	EndTime time.Time
+
+	Sources []string
+
+	CreatedAt time.Time
+}
+
+// TimelineEvent represents an event in a forensic timeline.
+
+type TimelineEvent struct {
+	ID string
+
+	Timestamp time.Time
+
+	Type string
+
+	Source string
+
+	Description string
+
+	Evidence []string
+
+	Significance string
+
+	Tags []string
+}
+
+// ForensicReporting provides forensic reporting.
+
+type ForensicReporting struct {
+	templates map[string]*ForensicReportTemplate
+
+	generators map[string]*ForensicReportGenerator
+
+	reports map[string]*ForensicReport
+}
+
+// ForensicReportTemplate defines forensic report templates.
+
+type ForensicReportTemplate struct {
+	ID string
+
+	Name string
+
+	Type string
+
+	Sections []*ReportSection
+
+	Format string
+
+	Template string
+}
+
+// ReportSection represents a section in a forensic report.
+
+type ReportSection struct {
+	ID string
+
+	Title string
+
+	Content string
+
+	Type string
+
+	Required bool
+
+	Order int
+}
+
+// ForensicReportGenerator generates forensic reports.
+
+type ForensicReportGenerator interface {
+	GenerateReport(ctx context.Context, template *ForensicReportTemplate, data interface{}) (*ForensicReport, error)
+
+	GetSupportedFormats() []string
+}
+
+// ForensicReport represents a generated forensic report.
+
+type ForensicReport struct {
+	ID string
+
+	IncidentID string
+
+	Type string
+
+	Title string
+
+	Summary string
+
+	Findings []*ForensicFinding
+
+	Evidence []string
+
+	Conclusions []string
+
+	Recommendations []string
+
+	Appendices []*ReportAppendix
+
+	GeneratedAt time.Time
+
+	GeneratedBy string
+
+	Content []byte
+
+	Format string
+}
+
+// ReportAppendix represents a report appendix.
+
+type ReportAppendix struct {
+	ID string
+
+	Title string
+
+	Type string
+
+	Content []byte
+
+	Filename string
+}
+
+// IncidentResponseConfig holds incident response configuration.
+
+type IncidentResponseConfig struct {
+	DefaultPlaybook string
+
+	AutoAssignments bool
+
+	EscalationEnabled bool
+
+	CommunicationChannels []string
+
+	ForensicsEnabled bool
+
+	DocumentationRequired bool
+
+	ReviewRequired bool
+
+	MetricsEnabled bool
+}
+
+// SecurityMetrics holds Prometheus metrics for security management.
+
+type SecurityMetrics struct {
+	AuthenticationAttempts *prometheus.CounterVec
+
+	AuthenticationFailures *prometheus.CounterVec
+
+	CertificatesManaged prometheus.Gauge
+
+	CertificatesExpiring prometheus.Gauge
+
+	SecurityAlerts *prometheus.CounterVec
+
+	ThreatDetections *prometheus.CounterVec
+
+	ComplianceScore *prometheus.GaugeVec
+
+	VulnerabilitiesFound *prometheus.CounterVec
+
+	IncidentsActive prometheus.Gauge
+
+	IncidentsResolved prometheus.Counter
+}
+
+// Configuration structures and additional types would continue...
+
+// For brevity, I'll include the main interface methods.
+
+// NewComprehensiveSecurityManager creates a new security manager.
+
+func NewComprehensiveSecurityManager(config *SecurityManagerConfig) *ComprehensiveSecurityManager {
+	if config == nil {
+		config = &SecurityManagerConfig{
+			AuthenticationMethods: []string{"password", "certificate"},
+
+			SessionTimeout: 30 * time.Minute,
+
+			MaxFailedAttempts: 3,
+
+			LockoutDuration: 15 * time.Minute,
+
+			AuditLogRetention: 90 * 24 * time.Hour,
+
+			ThreatDetectionEnabled: true,
+
+			IntrusionDetectionEnabled: true,
+
+			VulnerabilityScanInterval: 24 * time.Hour,
+		}
+	}
+
+	csm := &ComprehensiveSecurityManager{
+		config: config,
+
+		securityMetrics: initializeSecurityMetrics(),
+
+		stopChan: make(chan struct{}),
+	}
+
+	// Initialize all security components.
+
+	csm.certificateManager = NewCertificateLifecycleManager(config.CertificateAuthority)
+
+	csm.authenticationMgr = NewAuthenticationManager(&AuthenticationConfig{
+		SessionTimeout: config.SessionTimeout,
+
+		MaxFailedAttempts: config.MaxFailedAttempts,
+
+		LockoutDuration: config.LockoutDuration,
+	})
+
+	csm.authorizationMgr = NewAuthorizationManager()
+
+	csm.encryptionMgr = NewEncryptionManager(config.EncryptionStandards)
+
+	csm.keyManagementService = NewKeyManagementService()
+
+	csm.securityAuditMgr = NewSecurityAuditManager(config.AuditLogRetention)
+
+	csm.securityPolicyEngine = NewDefaultSecurityPolicyEngine()
+
+	csm.secureChannelMgr = NewSecureChannelManager(&SecureChannelConfig{})
+
+	if config.IntrusionDetectionEnabled {
+		csm.intrusionDetection = NewIntrusionDetectionSystem(&IDSConfig{})
+	}
+
+	if config.ThreatDetectionEnabled {
+		csm.threatDetectionMgr = NewThreatDetectionManager(&ThreatDetectionConfig{})
+	}
+
+	csm.complianceMonitor = NewComplianceMonitor(&ComplianceConfig{
+		EnabledFrameworks: config.ComplianceModes,
+	})
+
+	csm.vulnerabilityScanner = NewVulnerabilityScanner(&VulnerabilityScanConfig{
+		ScanSchedule: map[string]string{"daily": "02:00"},
+	})
+
+	csm.incidentResponseMgr = NewIncidentResponseManager(&IncidentResponseConfig{})
+
+	return csm
+}
+
+// Start starts the security manager.
+
+func (csm *ComprehensiveSecurityManager) Start(ctx context.Context) error {
+	csm.mutex.Lock()
+
+	defer csm.mutex.Unlock()
+
+	if csm.running {
+		return fmt.Errorf("security manager already running")
+	}
+
+	logger := log.FromContext(ctx)
+
+	logger.Info("starting comprehensive security manager")
+
+	// Start certificate lifecycle management.
+
+	if err := csm.certificateManager.Start(ctx); err != nil {
+		return fmt.Errorf("failed to start certificate manager: %w", err)
+	}
+
+	// Start intrusion detection.
+
+	if csm.intrusionDetection != nil {
+		if err := csm.intrusionDetection.Start(ctx); err != nil {
+			logger.Error(err, "failed to start intrusion detection")
+		}
+	}
+
+	// Start threat detection.
+
+	if csm.threatDetectionMgr != nil {
+		if err := csm.threatDetectionMgr.Start(ctx); err != nil {
+			logger.Error(err, "failed to start threat detection")
+		}
+	}
+
+	// Start vulnerability scanning.
+
+	if err := csm.vulnerabilityScanner.Start(ctx); err != nil {
+		logger.Error(err, "failed to start vulnerability scanner")
+	}
+
+	// Start compliance monitoring.
+
+	if err := csm.complianceMonitor.Start(ctx); err != nil {
+		logger.Error(err, "failed to start compliance monitor")
+	}
+
+	csm.running = true
+
+	logger.Info("comprehensive security manager started successfully")
+
+	return nil
+}
+
+// Stop stops the security manager.
+
+func (csm *ComprehensiveSecurityManager) Stop(ctx context.Context) error {
+	csm.mutex.Lock()
+
+	defer csm.mutex.Unlock()
+
+	if !csm.running {
+		return nil
+	}
+
+	logger := log.FromContext(ctx)
+
+	logger.Info("stopping comprehensive security manager")
+
+	close(csm.stopChan)
+
+	// Stop all components.
+
+	if csm.certificateManager != nil {
+		csm.certificateManager.Stop(ctx)
+	}
+
+	if csm.intrusionDetection != nil {
+		csm.intrusionDetection.Stop(ctx)
+	}
+
+	if csm.threatDetectionMgr != nil {
+		csm.threatDetectionMgr.Stop(ctx)
+	}
+
+	if csm.vulnerabilityScanner != nil {
+		csm.vulnerabilityScanner.Stop(ctx)
+	}
+
+	if csm.complianceMonitor != nil {
+		csm.complianceMonitor.Stop(ctx)
+	}
+
+	csm.running = false
+
+	logger.Info("comprehensive security manager stopped")
+
+	return nil
+}
+
+// Core security operations.
+
+// AuthenticateUser authenticates a user.
+
+func (csm *ComprehensiveSecurityManager) AuthenticateUser(ctx context.Context, credentials *AuthCredentials) (*AuthResult, error) {
+	return csm.authenticationMgr.Authenticate(ctx, credentials)
+}
+
+// AuthorizeAccess authorizes access to a resource.
+
+func (csm *ComprehensiveSecurityManager) AuthorizeAccess(ctx context.Context, request *AccessRequest) (*AccessDecision, error) {
+	return csm.authorizationMgr.Authorize(ctx, request)
+}
+
+// IssueCertificate issues a new certificate.
+
+func (csm *ComprehensiveSecurityManager) IssueCertificate(ctx context.Context, request *CertificateRequest) (*ManagedCertificate, error) {
+	return csm.certificateManager.IssueCertificate(ctx, request)
+}
+
+// GetSecurityStatus returns overall security status.
+
+func (csm *ComprehensiveSecurityManager) GetSecurityStatus(ctx context.Context) (*SecurityStatus, error) {
+	// Implementation would aggregate status from all security components.
+
+	status := &SecurityStatus{
+		ComplianceLevel: "HIGH",
+
+		ActiveThreats: []string{},
+
+		LastAudit: time.Now().Add(-24 * time.Hour),
+
+		Metrics: json.RawMessage(`{}`),
+	}
+
+	return status, nil
+}
+
+// Helper methods and placeholder implementations.
+
+func (csm *ComprehensiveSecurityManager) getActiveAlertCount() int {
+	// Would aggregate from all security components.
+
+	return 0
+}
+
+func initializeSecurityMetrics() *SecurityMetrics {
+	return &SecurityMetrics{
+		AuthenticationAttempts: promauto.NewCounterVec(prometheus.CounterOpts{
+			Name: "oran_security_auth_attempts_total",
+
+			Help: "Total number of authentication attempts",
+		}, []string{"method", "result"}),
+
+		AuthenticationFailures: promauto.NewCounterVec(prometheus.CounterOpts{
+			Name: "oran_security_auth_failures_total",
+
+			Help: "Total number of authentication failures",
+		}, []string{"method", "reason"}),
+
+		CertificatesManaged: promauto.NewGauge(prometheus.GaugeOpts{
+			Name: "oran_security_certificates_managed",
+
+			Help: "Number of certificates under management",
+		}),
+
+		CertificatesExpiring: promauto.NewGauge(prometheus.GaugeOpts{
+			Name: "oran_security_certificates_expiring",
+
+			Help: "Number of certificates expiring soon",
+		}),
+
+		SecurityAlerts: promauto.NewCounterVec(prometheus.CounterOpts{
+			Name: "oran_security_alerts_total",
+
+			Help: "Total number of security alerts",
+		}, []string{"type", "severity"}),
+
+		ThreatDetections: promauto.NewCounterVec(prometheus.CounterOpts{
+			Name: "oran_security_threats_detected_total",
+
+			Help: "Total number of threats detected",
+		}, []string{"type", "severity"}),
+
+		ComplianceScore: promauto.NewGaugeVec(prometheus.GaugeOpts{
+			Name: "oran_security_compliance_score",
+
+			Help: "Compliance score by framework",
+		}, []string{"framework"}),
+
+		VulnerabilitiesFound: promauto.NewCounterVec(prometheus.CounterOpts{
+			Name: "oran_security_vulnerabilities_found_total",
+
+			Help: "Total number of vulnerabilities found",
+		}, []string{"severity", "type"}),
+
+		IncidentsActive: promauto.NewGauge(prometheus.GaugeOpts{
+			Name: "oran_security_incidents_active",
+
+			Help: "Number of active security incidents",
+		}),
+
+		IncidentsResolved: promauto.NewCounter(prometheus.CounterOpts{
+			Name: "oran_security_incidents_resolved_total",
+
+			Help: "Total number of resolved security incidents",
+		}),
+	}
+}
+
+// Placeholder implementations for major security components.
+
+// In production, each would be fully implemented with comprehensive functionality.
+
+// CertificateRequest represents a certificate issuance request.
+
+type CertificateRequest struct {
+	Subject pkix.Name
+
+	AlternativeNames []string
+
+	KeySize int
+
+	Usage []string
+
+	ValidityPeriod time.Duration
+
+	Template string
+}
+
+// NewCertificateLifecycleManager performs newcertificatelifecyclemanager operation.
+
+func NewCertificateLifecycleManager(caConfig *CAConfig) *CertificateLifecycleManager {
+	return &CertificateLifecycleManager{
+		certificates: make(map[string]*ManagedCertificate),
+
+		autoRenewal: &AutoRenewalService{},
+
+		validationService: &CertificateValidationService{},
+	}
+}
+
+// Start performs start operation.
+
+func (clm *CertificateLifecycleManager) Start(ctx context.Context) error { return nil }
+
+// Stop performs stop operation.
+
+func (clm *CertificateLifecycleManager) Stop(ctx context.Context) error { return nil }
+
+// IssueCertificate performs issuecertificate operation.
+
+func (clm *CertificateLifecycleManager) IssueCertificate(ctx context.Context, request *CertificateRequest) (*ManagedCertificate, error) {
+	return &ManagedCertificate{}, nil
+}
+
+// GetCertificateCount performs getcertificatecount operation.
+
+func (clm *CertificateLifecycleManager) GetCertificateCount() int { return len(clm.certificates) }
+
+// NewAuthenticationManager performs newauthenticationmanager operation.
+
+func NewAuthenticationManager(config *AuthenticationConfig) *AuthenticationManager {
+	return &AuthenticationManager{
+		methods: make(map[string]AuthenticationMethod),
+
+		sessions: make(map[string]*AuthSession),
+
+		failedAttempts: make(map[string]*FailedAttemptTracker),
+
+		config: config,
+	}
+}
+
+// Authenticate performs authenticate operation.
+
+func (am *AuthenticationManager) Authenticate(ctx context.Context, credentials *AuthCredentials) (*AuthResult, error) {
+	return &AuthResult{Success: true}, nil
+}
+
+// GetActiveSessions performs getactivesessions operation.
+
+func (am *AuthenticationManager) GetActiveSessions() int { return len(am.sessions) }
+
+// NewAuthorizationManager performs newauthorizationmanager operation.
+
+func NewAuthorizationManager() *AuthorizationManager {
+	return &AuthorizationManager{
+		rbacEngine: &RoleBasedAccessControl{},
+
+		abacEngine: &AttributeBasedAccessControl{},
+
+		policyEngine: &PolicyEngine{},
+
+		accessDecisions: make(map[string]*AccessDecision),
+	}
+}
+
+// Authorize performs authorize operation.
+
+func (azm *AuthorizationManager) Authorize(ctx context.Context, request *AccessRequest) (*AccessDecision, error) {
+	return &AccessDecision{Decision: "PERMIT"}, nil
+}
+
+// NewEncryptionManager performs newencryptionmanager operation.
+
+func NewEncryptionManager(config *EncryptionConfig) *EncryptionManager {
+	return &EncryptionManager{
+		encryptors: make(map[string]Encryptor),
+
+		config: config,
+	}
+}
+
+// NewKeyManagementService performs newkeymanagementservice operation.
+
+func NewKeyManagementService() *KeyManagementService {
+	return &KeyManagementService{
+		keys: make(map[string]*CryptographicKey),
+
+		keyPolicies: make(map[string]*KeyPolicy),
+	}
+}
+
+// NewSecurityAuditManager performs newsecurityauditmanager operation.
+
+func NewSecurityAuditManager(retention time.Duration) *SecurityAuditManager {
+	return &SecurityAuditManager{
+		auditLog: &AuditLog{entries: make([]*AuditEntry, 0)},
+
+		auditPolicies: make([]*AuditPolicy, 0),
+	}
+}
+
+// NewDefaultSecurityPolicyEngine performs newdefaultsecuritypolicyengine operation.
+
+func NewDefaultSecurityPolicyEngine() security.SecurityPolicyEngine {
+	return &DefaultSecurityPolicyEngine{
+		policies: make(map[string]*security.SecurityPolicy),
+
+		evaluators: make([]PolicyEvaluator, 0),
+
+		cache: &PolicyCache{decisions: make(map[string]*CachedDecision)},
+	}
+}
+
+// DefaultSecurityPolicyEngine provides a default implementation of SecurityPolicyEngine.
+
+type DefaultSecurityPolicyEngine struct {
+	policies map[string]*security.SecurityPolicy
+
+	evaluators []PolicyEvaluator
+
+	cache *PolicyCache
+
+	mutex sync.RWMutex
+}
+
+// EvaluatePolicy performs evaluatepolicy operation.
+
+func (dpe *DefaultSecurityPolicyEngine) EvaluatePolicy(request *security.PolicyRequest) (*security.PolicyDecision, error) {
+	// Simplified implementation for compilation.
+
+	return &security.PolicyDecision{
+		Decision: "PERMIT",
+
+		Reason: "Default allow policy",
+	}, nil
+}
+
+// AddPolicy performs addpolicy operation.
+
+func (dpe *DefaultSecurityPolicyEngine) AddPolicy(policy *security.SecurityPolicy) error {
+	dpe.mutex.Lock()
+
+	defer dpe.mutex.Unlock()
+
+	dpe.policies[policy.ID] = policy
+
+	return nil
+}
+
+// RemovePolicy performs removepolicy operation.
+
+func (dpe *DefaultSecurityPolicyEngine) RemovePolicy(policyID string) error {
+	dpe.mutex.Lock()
+
+	defer dpe.mutex.Unlock()
+
+	delete(dpe.policies, policyID)
+
+	return nil
+}
+
+// UpdatePolicy performs updatepolicy operation.
+
+func (dpe *DefaultSecurityPolicyEngine) UpdatePolicy(policy *security.SecurityPolicy) error {
+	dpe.mutex.Lock()
+
+	defer dpe.mutex.Unlock()
+
+	dpe.policies[policy.ID] = policy
+
+	return nil
+}
+
+// GetPolicy performs getpolicy operation.
+
+func (dpe *DefaultSecurityPolicyEngine) GetPolicy(policyID string) (*security.SecurityPolicy, error) {
+	dpe.mutex.RLock()
+
+	defer dpe.mutex.RUnlock()
+
+	if policy, exists := dpe.policies[policyID]; exists {
+		return policy, nil
+	}
+
+	return nil, fmt.Errorf("policy not found: %s", policyID)
+}
+
+// ListPolicies performs listpolicies operation.
+
+func (dpe *DefaultSecurityPolicyEngine) ListPolicies(filter *security.PolicyFilter) ([]*security.SecurityPolicy, error) {
+	dpe.mutex.RLock()
+
+	defer dpe.mutex.RUnlock()
+
+	var result []*security.SecurityPolicy
+
+	for _, policy := range dpe.policies {
+		result = append(result, policy)
+	}
+
+	return result, nil
+}
+
+// NewSecureChannelManager performs newsecurechannelmanager operation.
+
+func NewSecureChannelManager(config *SecureChannelConfig) *SecureChannelManager {
+	return &SecureChannelManager{
+		channels: make(map[string]*SecureChannel),
+
+		config: config,
+	}
+}
+
+// NewIntrusionDetectionSystem performs newintrusiondetectionsystem operation.
+
+func NewIntrusionDetectionSystem(config *IDSConfig) *IntrusionDetectionSystem {
+	return &IntrusionDetectionSystem{
+		sensors: make(map[string]*IDSSensor),
+
+		rules: make([]*IDSRule, 0),
+
+		alerts: make(chan *SecurityAlert, 1000),
+
+		config: config,
+	}
+}
+
+// Start performs start operation.
+
+func (ids *IntrusionDetectionSystem) Start(ctx context.Context) error { ids.running = true; return nil }
+
+// Stop performs stop operation.
+
+func (ids *IntrusionDetectionSystem) Stop(ctx context.Context) error { ids.running = false; return nil }
+
+// NewThreatDetectionManager performs newthreatdetectionmanager operation.
+
+func NewThreatDetectionManager(config *ThreatDetectionConfig) *ThreatDetectionManager {
+	return &ThreatDetectionManager{
+		detectors: make(map[string]*ThreatDetector),
+
+		threatDatabase: &ThreatDatabase{},
+
+		config: config,
+	}
+}
+
+// Start performs start operation.
+
+func (tdm *ThreatDetectionManager) Start(ctx context.Context) error { tdm.running = true; return nil }
+
+// Stop performs stop operation.
+
+func (tdm *ThreatDetectionManager) Stop(ctx context.Context) error { tdm.running = false; return nil }
+
+// NewComplianceMonitor performs newcompliancemonitor operation.
+
+func NewComplianceMonitor(config *ComplianceConfig) *ComplianceMonitor {
+	return &ComplianceMonitor{
+		frameworks: make(map[string]*ComplianceFramework),
+
+		assessments: make(map[string]*ComplianceAssessment),
+
+		controls: make(map[string]*ComplianceControl),
+
+		config: config,
+	}
+}
+
+// Start performs start operation.
+
+func (cm *ComplianceMonitor) Start(ctx context.Context) error { return nil }
+
+// Stop performs stop operation.
+
+func (cm *ComplianceMonitor) Stop(ctx context.Context) error { return nil }
+
+// GetOverallScore performs getoverallscore operation.
+
+func (cm *ComplianceMonitor) GetOverallScore() float64 { return 85.5 }
+
+// NewVulnerabilityScanner performs newvulnerabilityscanner operation.
+
+func NewVulnerabilityScanner(config *VulnerabilityScanConfig) *VulnerabilityScanner {
+	return &VulnerabilityScanner{
+		scanners: make(map[string]*VulnScanner),
+
+		database: &VulnerabilityDatabase{},
+
+		config: config,
+	}
+}
+
+// Start performs start operation.
+
+func (vs *VulnerabilityScanner) Start(ctx context.Context) error { vs.running = true; return nil }
+
+// Stop performs stop operation.
+
+func (vs *VulnerabilityScanner) Stop(ctx context.Context) error { vs.running = false; return nil }
+
+// GetVulnerabilityCount performs getvulnerabilitycount operation.
+
+func (vs *VulnerabilityScanner) GetVulnerabilityCount() int { return 0 }
+
+// NewIncidentResponseManager performs newincidentresponsemanager operation.
+
+func NewIncidentResponseManager(config *IncidentResponseConfig) *IncidentResponseManager {
+	return &IncidentResponseManager{
+		incidents: make(map[string]*SecurityIncident),
+
+		playbooks: make(map[string]*IncidentPlaybook),
+
+		responders: make(map[string]*IncidentResponder),
+
+		workflows: make(map[string]*ResponseWorkflow),
+
+		config: config,
+	}
+}
+
+// Additional configuration types.
+
+type SecureChannelConfig struct {
+	DefaultTLSVersion uint16
+
+	CipherSuites []uint16
+
+	MaxConnections int
+
+	IdleTimeout time.Duration
+}
+
